Fix Android Studio setup for support lib
There seems to be an issue with MediaRouter and
AppCompat with the API moduel setup.

MR depends on AppCompat which contains resources
which need the current SDK. Android Studio can't
handle this and tries to compile AppCompat with the
ApiModule's SDK version which then breaks
(AppCompat can't be compiled against 16, etc).
To work around this, we compile the ApiModules against
current when we're using AS only.

Change-Id: I1a4dbbb95456f40067524de43e522fe6ef704a33


GitOrigin-RevId: 1eff297d6cc33c6a55530c193f87eb5b07da2562