Bring back <type>aar</type> hack.
Some of our external clients relied of <type>aar</type> tag
inside the <dependencies> list in the POMs of support libraries.
Ideally, Android Gradle Plugin would handle this for us, but it does
not, so as a workaround add this hack. Note, this will only work for
support library dependencies (not architecture components).

Bug: 67099985
Test: ./gradlew createArchive
Change-Id: Ib854a0e7d9efd6dbbdc16e35961525a1b5c0ea8c


GitOrigin-RevId: eb45a18cbde2a90e4ba70b822439ac115b39925a