Read kotlin parameter names from Kotlin Metadata
This is a "best effort" CL to read kotlin metadata in classes using a
3rd party library. They are technically part of kotlin API so safe
to read but JB does not provide a library so there is no future
compatibility guaratees (though it uses protobuf so needs to be somewhat
backwards compatible).

This CL uses the library under paranoid checks so that if something
goes wrong, normal compilation wont fail, only this feature will.

Bug: 67181813
Test: KotlinTestApp#DependecyDaoTest
Change-Id: I5316b0ecb04addd22a6e881024036106124c001a


GitOrigin-RevId: 9e828633846e3c3ff9738dbc567fe16c05e5d1b4