Add a transaction like data frame
A data frame implements an multi-version concurrency control algorithm
to implement a transaction like frame that allows objects using it to
support atomic, consistent, and isolated changes. The objects are not
durable and the changes are not serializable consistent (as reads
crossing writes between frames are not disallowed).

Frames can be opened as read-only which provides an immutable view of
the objects (all writes are isolated and the frame disallows changes).

Change-Id: I7537c36f7ea042c8d3c676ab58523ca8be509ee7


GitOrigin-RevId: a72fe6ef8432e8b1f85c0e9bf23e67153da3ffbc