Fix jetifier zip generation on Windows.
The bug was caused by putting Windows paths into the zip stream which
generated a zip file that was observed as flatten by the tooling.
Interestingly enough it was not seen as flatten by Windows itself.

Also make sure we rewrite path of a file everytime not only when
something was changed inside it. That was not intended behavior.

Plus bump versions of SL to alpha3.

Minor change: ignore new ktx stuff from dejetification.

Fix: 109738608
Bug: 64833662

Test: Manually verified that running jetifier on Windows produces
correct zip file after the fix.

Change-Id: Ib10d47a74afdd4b462309294634c429ac0bc1bc0


GitOrigin-RevId: dbd0259de831cc05624ed9b7e5930d8d106631d7