Convert XML to KTX code when copy pasting XML in Kotlin file.
This CL also adds a "Convert XML File to Component" option when right clicking an XML file on the file tree.

I added a lot of TODOs to keep track of what should be improved. Most important one is the mapping from XML attribute (name and value) to KTX attribute.

Change-Id: I82bf0f624d527bd9c5f7ca74ed84e2ee0f57fe13


GitOrigin-RevId: b07ae1b381dd2e7ee6e87128efcc2debc4f8b1c3