Generalize lockClocks.sh to support all rooted devices
Bug: 111923337
Test: manually tested on Bullhead, Marlin, Walleye

Generalize lockClocks to inspect frequency list, and locks (what it
guesses to be) the biggest cores to ~50% of max, disabling others.

Also verify that written clock values succeed.

Removed set -e, since bullhead for some reason errors with each of
its writes to cpufreq, but we manually verify that the writes succeed,
so *shrug*.

Change-Id: I703de6e8675b084f5c80a3bb1e960cbbf15e3c4d


GitOrigin-RevId: b3dcc34ed27bb6831b1df9a02c03259f72c4728e