Reuse overlay view between two Visibility transitions
Bug: 31362791

Instead of creating a new screenshot of the view every time previous visibility transition already did it or just added our startView to the overlay we can cache it and apply the animation to the same reused view. So the behaviour is the same(and expected) like in the case when no overlay is used.

Test: test added in VisibilityTest
Change-Id: I5fe548de2a8de7f803b93fde46592c88e9af47eb


GitOrigin-RevId: 66b75fbcf2e5baf1439f862b82bc643f0991f02f