Support for verifying dependency versions
Added a task to android and java library projects that enables
them to test that their dependency versions are compatible with
their versions.

A project is compatible with a particular dependency if the
dependency version is of an equal or later release phase. This
means that a beta project can only have beta, rc or stable
dependencies, an rc project can only have rc and stable
dependencies and stable projects can only have stable
dependencies.

Eventually buildOnServer will depend on this task to enforce
this rule however given that some libraries do not currently
abide by this rule yet, this task will need to be ran manually.

Test: run verifyDependencyVersions on a particular project while
tweaking the project's version and the dependency versions and
ensure expected behaviour.

Change-Id: Ic7778fdf08704cc78e73cda40f27ca90f3295030


GitOrigin-RevId: 40913676a96590a0676921fafabc3d5f70fe1555