Defines GUAVA_LISTENABLE_FUTURE_AVOID_CONFLICT
Defines GUAVA_LISTENABLE_FUTURE_AVOID_CONFLICT to point to the empty
definition of ListenableFuture, which can be used to avoid conflicts
with guava. go/aog/757269 is adding GUAVA_LISTENABLE_FUTURE to the
dependencies of |core| module, and some other modules are using both
core and TRUTH(which pulls in Guava), which was causing multiple
dependency conflicts.

Test: None
Change-Id: I0f1fee36f4e14ac2e1176f8c6cc8e9fe7c26149c


GitOrigin-RevId: ce15356fc9dfa98db0d18ce961c7ed2dccc6b37f