Improve conversion on pasting + add tests.
This CL adds tests for the conversion performed when converting XML to
KTX code when pasting XML into a Kotlin file.

The pasted KTX code is now automatically wrapped in a @Composable
function if pasted outside a callable.

When pasted inside a function that doesn't have the annotation, we
suggest the user to add it (this can be enabled automatically, without
prompt).

Change-Id: I0bb0a337167266a40cca8a20c3117cf348b76a05


GitOrigin-RevId: a16102a9adf9bbb11fddbefa98a80163feac0b18