[R4A] Allow model object to be created during static initialization
With this change model object can be created (but not modified or
read) during static initialization.

If a model object requires mutatation or read during initialization
(such as creating a circular reference) the statement or expression
can be surrounded by `framed { ... }` which will create a temporary
frame around the expression or statement.

Fixes: b/120843442
Change-Id: I247c7ca8a99773f36b575056337d1f10cc7f7025


GitOrigin-RevId: 45d8c06854f5a2c9cd6624f6591f80209adb9dea