Avoid eager evaluation due to docs
This CL fixes an issue in DiffAndDocs where it wants to make sure other
projects are evaluated before diff and docs yet this causes gradle to
eagerly evaluate all other projects. Now we do that setup right before
docs is evaluated so that if docs will not be evaluated, it won't trigger
evaluation of all others.

Bug: 121135187
Test: gw anchorDocsTask -PofflineDocs=true
Change-Id: I9b2a069f6a563da39faafd838d385e795812c98a


GitOrigin-RevId: cfa17144195d41819e807bad9b7fe204b3304f49