Fixed broken min/max testing
Lazy evaluation caused maxDepVersions tasks to not switch
dependencies due to module:project map not being complete.

Due to test application ids being similar for all flavors
(apparently an AGP bug), minDepVersions and maxDepVersions
tests could not be ran simultaneously on a device, which is
also a problem.

The following CL forces evaluation of all children before
dependency substitution and substitutes for the debug
variant (cannot substitute for release variant after
evaluation.)

Removed flavors completely to make the build faster since
it makes more sense to test binary compatibility during
presubmits/postsubmits. We now use a flag
(-PuseMaxDepVersions) instead which triggers substitution.
Only debug configurations/tasks are substituted for, which
is fine since buildOnServer, buildTestApks build using
debug tasks. Lint also runs on the debug variant so this
works for our purposes (although it is currently disabled
for maxdepversions as I am investigating some behavior.)

The plan is to run presubmit and postsubmit targets twice
with every second target running with the -PuseMaxDepVersions
flag. This parallelizes runs and guarantees binary compatibility
testing while still running in theoratically the same time as
before flavors (although we are likely to use more resources.)

Test: ./gradlew <project>:dependencies -PuseMaxDepVersions

Observe the dependencies change to project dependencies in debug
configurations. (compared to a run without the flag.)
Also deprecate a method in ToT and make sure that using the flag
on a project depending on a released artifact of the project
containing the deprecated class show a warning about using
deprecated class meaning it did indeed switch to the ToT version
of that dependency. (introducing an error is also another method
of testing, although I personally have not tried it this time.)

Change-Id: Ic9c7a3c5844cda20f1d7e0acd6cf856b6fe035c1


GitOrigin-RevId: 9ff667a9247b52cac5667ef10dcbbc703ab2cee6