Recompose should not try to recreated component that are unmodified
When processing nested recomposes, the stack of groups used for
recomposition was confused by recompose scopes that ended at the same
location in the slot table causing a group to "left" prematurely. This
caused a component to be deleted and inserted incorrectly causing
in some cases, the wrong item to be removed.

Fixes: b/123348629
Change-Id: I994a518c42e876145ac99812673b5d45b347c01f


GitOrigin-RevId: 0cf0524348c7618ef88c1a92ca5ac957ef357ee2