Enable building of co-built sets
Per b/128577735, media team has tests that are dependent on
one another, but their modules are not dependent. With our testing
strategy of partitioning the build and testing each part separately
(between the dependent modules and the directly changed modules)
we have been causing infra errors in their pre-submits.

This change makes it so that if a project with a codependent test
is built, the other modules will also be built to avoid the error.

It is general enough to enable us to easily add more cobuilt modules
in the future should the need arise for another team/situation.

Change-Id: I9bfcce387a7aca8d59ff5c3bfcbb6bae2c5c5fc9
Fix: 128577735
Test: Added more unit tests for this special case.


GitOrigin-RevId: 6d9ae7d7d35449f3bc98b133684edaffa2c70d86