Activity launching benchmark runner
Bug:111923337
Bug:129282573

Adds a (currently optional) benchmark runner to the benchmark library, which when used launches an
'isolation activity' to ensure our benchmarks are running on top, without the launcher or sysui
showing.

Currently, there's no label on the benchmark activity, so device inexplicably blinks black.

Additionally, we don't yet resume our activity if a benchmark launches an activity, e.g. via a
ActivityTestRule or ActivityScenario. This will be done in a follow up.

This work is a prerequisite to using Window.setSustainedPerformanceMode on devices that support it.

Test: AndroidBenchmarkRunnerTest

Change-Id: Id9dd5471ebd02b0639e46ee450568661b8da5f2e


GitOrigin-RevId: 5412b2cf3b7ec5c53512f6b5c5e22ec70cd6e319