Remove targetSdkVersion workaround for library tests
Turns out that manifest merger picks the lowest target sdk version,
so it is fine to set the targetSdkVersion directly on our libraries.

https://developer.android.com/studio/build/manifest-merge

Also turns out lint was not running all the rules when we were not
setting targetSdkVersion, so fix up all the existing uses except for 2
that have followup bugs.

Bug: 132905855,132906560
Test: ./gradlew createArchive -> now our aar AndroidManifest.xml has
targetSdkVersion, but that's ok as manifest merger handles this
correctly.

Change-Id: I25c3a30c282888659e9b3177f1f9a0014c6a9a56


GitOrigin-RevId: a1aa570607371188c68469577de50a2505d27eb7