Add benchmarks module.
The module is child of integration-tests module which is mainly meant to
contain correctness tests (such as asserting metrics counts) to ensure
validity of the benchmarks itself.

The idea is that each benchmark would have a TestCase class containing
most of the logic and we would run it twice. Once in benchmarks and
second time in the integration tests. In the integration tests we can do
variaty of asserts including easier debugging.

I have also added a boolean to recomposeSync so we know if there were
any changes done during that. This helps a lot with correctness tests.

Test: Added
Bug: 127479944
Change-Id: I2e24affb695ef04c5e9814c5db97ccfa8d5a1241


GitOrigin-RevId: 204479008dd4ef947b486ea09ca6e70215e3d498