Test to validate a fix for  crash when a invalidation arrives late.
If an invalidation arrived for a part of the composition
was removed, it would cause a -1 to be entered into the
invalidations list of the composer causing the next
recompose to crash. This was fixed in a previous commmit.
This commit adds a validation test.

Also fixes an issue with animations that would crash if it
complete after the component being animated was removed from
the composition.

Test: New test, ./gradle test
BUG: 137284953

Change-Id: Ib559b181069db1b24487b1709664fd48848d5136


GitOrigin-RevId: 0c6e0f1d36fc22d0964b1f78d38d9bca57d1632c