Merge android tests back into compose-runtime
Instrumented device tests should live in androidAndroidTest

Also adds multiplatform specific configuration to AndroidXPlugin,
namely:

Moves manifest files to androidMain and androidAndroidTest for
consistency with the android-specific modules

Adds multiplatform sourcesets to JavaCompileInputs so metalava and dokka
will work correctly on multiplatform modules

Android lint is still broken, and API tracking hasn't been turned on yet
as there are some issues in metalava that need to be fixed first to do
with false alarms.

Also, as there is no androidTest source set, running
connectedAndroidTest from the command line doesn't do anything - this
shouldn't affect presubmits as we manually instrument the built test
apks, but we should fix this in the future - one possible workaround is
adding an empty source file to src/androidTest.

Bug: 138741704
Bug: 138720404
Bug: 139276849
Test: compose-runtime tests
Change-Id: Ib6367c22ccbb15ab8511306cb366f1794dd553b4


GitOrigin-RevId: 8a1572411435ac68589485b7b5d2d9f9ba3209a6