Set requireSameVersion bit for media2
This CL sets requireSameVersion bit for the media2 library group
to synchronize the version of each libraries from 1.1.0. To do that
it bumps versions of media2-exoplayer and media2-widget to 1.1.0-alpha01.

Also, it removes the dependency on concurrent-futures from media2-widget
as it doesn't seem to be necessary.

Fix: 139265165
Test: ./gradlew checkApi verifyDependencyVersions &&
  ./gradlew :media2:media2-widget:connectedCheck
Change-Id: I919a7e0369b1a23bab404a831d4b565800cdf804


GitOrigin-RevId: 74fe431aceaa1b4e7f6f96fcfb51083dbda328eb