Ensure that override configuration is always a diff of the base config
Otherwise, we'll be stuck forever using the exact configuration that was
used as the initial override.

Add tests for maintaining locale overrides

Removes explicit locale change in AppCompat's BaseTestActivity, which
was interfering with the ability to override the locale. Also ensures
custom configuration only runs on API level 17+ since that's when the
custom configuration API was added.

Disables test on API 26 due to issues with setRequestedOrientation.

Bug: 138325067
Bug: 140503890
Bug: 141009022
Bug: 140607881
Fixes: 141457323
Test: NightModeOrientationConfigChangesTestCase
Test: NightModeCustomConfigurationTestCase
Change-Id: I2066b7cf67d2f72f16a44d5194b2fa89ca29ec63


GitOrigin-RevId: 2e75cdbabd18ad1d9fc267f833ebe9d3fdbf471b