Adds new error for composable type inference mismatches
The default error for type inference mismatch (from Kotlin) uses
RENDER_TYPE as the type renderer, however this does not preserve
annotations. Given that the type of a composable only comes from the
annotation, instead we need to use RENDER_TYPE_WITH_ANNOTATIONS which
correctly renders the annotation in the error message.

For example, trying to pass a composable lambda to a non-composable
parameter will print the following:

Type inference failed. Expected type mismatch:
inferred type is @Composable() () -> Unit but () -> Unit was expected

Bug: b/142756333
Test: ComposableCheckerTests
Change-Id: I34d9e608ff07a85cdea0df8bddfbf78dbe67ad67


GitOrigin-RevId: 978f78a48648f8ad2fb19f4bf293c39dd5ecb5d3