Model reads during measure stage should survive Frame switches
Previously we were calling currentFrame().observeReads(block) which wasn't working for some tricky cases when within this block we switch the Frame: the observer was still connected to the previous not-active frame and wasn't notified about the model reads. This issue was reproducible when inside WithConstraints we call DrawVector which is causing an additional subcomposition which is switching to a new frame.
To fix this I introduced global thread local list of read observers which can survive frame switches. And we fully control when they are added and removed as our api only provides observeReads(block) api, so the observer can't be leaked.

Bug: 144493391
Test: the fix is covered with tests on three different layers.
Change-Id: Ibba72c3f791c1435390ce12abfedac70729a5071


GitOrigin-RevId: 719cde6ee79f09c631e616400e61c649be32d320