Enables incremental Kotlin compilation in Compose projects
This CL also allows using a compiler daemon, which should greatly increase the
speed of compilations.

A couple of issues were noted which this cl works around:

1. Inline override functions appear to break when:
a) in incremental / daemon mode
b) using IR for compilation

In any case Kotlin compiler will throw a warning for doing this, so we
probably shouldn't do this anyway.

2. I also noticed a strange issue where ComposableAnnotationChecker.get()
throws an exception, because the plugin appears to not have been
registered. It looks like it should be safe just to create a new one
anyway, but it seems like there might be some strange case when using
the daemon, where the project instance is stale or there is a race
condition in adding the plugins to the project.

Bug: b/144304390
Test: ./gradlew bOS, a few incremental compilations
Change-Id: Ia001c864199ba6d7879952ede8194b6f1b97cad3


GitOrigin-RevId: 346b0896e3726d77e4fc22a519eb3c941916db14