Updated ArrowDrawable navigation-ui proguard rule
The proguard rule to keep the DrawerArrowDrawable class in minified
release builds was written before Navigation was moved to androidx. That
means for it to work on Navigation release that are in androidx,
jetifier needs to be enabled for the project. Now that navigation has
moved, we should adjust the rule so that jetifier is not required it to
work.

Test: Tested in sample app
Bug: 147610424
Change-Id: Iaac0c6aab4aa6340decbeeed2967a91133fe02b6
(cherry picked from commit 569ef82c1c7ec719efd4811d28e7801967efffc0)