Ensure onInflate is always called before Fragment is attached
With the changes in aosp/1183079 fragments added with the <fragment> tag
but not in a layout were allowed to move to CREATED so that they could
be cleaned up appropriately. As a side affect, fragments not in layout
could now be CREATED by the time they get the onInflate() callback. This
could cause unexpected behavior for users who expect onInflate() to
always be called before the fragment is ATTACHED.

This change ensures that fragments added with the <fragment> tag, but
not in a layout do not move to created until its fragmentManager is at
least ACTIVITY_CREATED. This ensures we continue to call onInflate()
before the fragment is ATTACHED, but also that inflated fragments are
cleaned up appropriately.

Test: Modified StrictFragment
Fixes: 147462530
Change-Id: Id9935535d8394cba22d201c8bc6fd63638d0a1d6


GitOrigin-RevId: 0bc25bbc8c161cdadfd1217869d8294f8ad87700