Remove Draw tag usages in layout, framework, material, benchmarks
This CL is a step to Draw tag deprecation and removal. It removes Draw tag usages in material, framework, demos, Rally and layout files.

As a side effect of this, SimpleComposables.kt that containst custon Container, ColoredRects and alighment in gesture demos was replaced with public API composables (modifiers, Box, etc.)

Bug: 149827027
Test: existent tests should pass
Change-Id: I399b8a72d76427884ad2d3879a343a4dd38fa5e7


GitOrigin-RevId: 0e4dc6d38359fbad83c7d21400a8217676213042