Separate window extension interfaces into different packages
The extension interface has to be updated to
androidx.window.extensions. However, an older version v01 was shipped
on some devices with androidx.window.sidecar package name and
slightly different interfaces. This CL separates the old and the
current versions into different modules, both of which are declared
as dependencies for the core module. To handle these different
version a new wrapper and abstraction was introduced in the core
module, and the ExtensionWindowBackend is provided with the correct
wrapper for available OEM implementation.

Test: androidx.window.ExtensionTest
Test: Manual, using sample on foldable device
Change-Id: I6f7c38512c63070fa97c910435ee41804041496f


GitOrigin-RevId: 8d094948e1c29a28d106719e50ad0f0f7e3c4207