Shadow Room's antlr dependency
Anltr runtime does a version check between itself and generated
code and might print a warning if they don't match.
To be safe, Room compiler jarjar's its Antlr dependency so that
it cannot conflict with other annotation processors which might
also use antlr (e.g. data binding)

To ensure it does not break, I've added a verification task to
Room compiler module and also added it as a dependeny to the
buildOnServer task.

The task checks the generated pom for room-compiler to make sure
it does not include antlr but it includes at least one another
dependency (kotlin-stdlib).

I've also added checks for the room-compiler-jar contents to double
check antlr is there and its package is moved.

Bug: 150106190
Test: buildOnServer verifies the output

Change-Id: I2c99cba0b3b42369a4b495e5c086f18f5d0b3bff


GitOrigin-RevId: 562a37df135f25e77f9103f695308d3945c2e2c7