Updated DrawScope and ContentDrawScope to extend from CanvasScope to access the declarative graphics framework
Relnote: "Updated higher level
compose APIs that expose a Canvas to
expose CanvasScope instead. This removes
the need for consumers to maintain their
own Paint objects. For consumers that
still require access to a Canvas
they can use the drawCanvas extension
method which provides a callback to issue
drawing commands with the underlying
Canvas."

Bug: 154835459
Test: re-ran compose tests
Change-Id: I80afdf4c0a648962aa6ef1efc05b1d3b65757094


GitOrigin-RevId: dcb72209f8caa74d6f379b7c3f288e993351801e