Introduce receiver scope for WithConstraints
This allows to simplify the usage as you will not need to write something like:
    WithConstraints { constraints, _ -> }
When you need constraints, but not layout direction.
Plus allows us to provide additional helper properties minWidth, maxWidth, minHeight and maxHeight which are values in Dp instead of pixels. This simplifies the main usecase of the WithConstraints were you need to check if the available size in dp is within the threshold you have in Dp.

Relnote: WithConstraints trailing lambda API has been changed. Now instead of two params it has a receiver scope which in addition to constraints and layoutDirection provides minWidth, maxWidth, minHeight and maxHeight properties in Dp
Fixes: 149979702
Test: new tests
Change-Id: I91b9af740cd2613ddd4fe6d63cd539a46b52fc52


GitOrigin-RevId: 0f072e5cb6523653474e7e9e1fa71180299b0821