Throw IllegalStateException on re-using keys
PageFetcherSnapshot now throws if it detects key re-use on attempts to
load distinct pages. This means re-using keys for loading a dropped page
is allowed and that key-reuse state is not carried over between
generations.

Adds a property to PagingSource to turn off this behavior, for
intentional use cases. In particular, this check was added to cover the
common case of infinitely looping empty loads as some item-keyed sources
and paging2 datasource marks end of load with an empty page, but an
empty load is not a terminal state in paging3.

Fixes: 156110590
Test: ./gradlew paging:paging-common:test
Change-Id: I0229d259f3ed4b8cf2bb1dda5cc2cd6014f59cbc


GitOrigin-RevId: 393212124606d2990953331bdf14648f0d00473e