Always verifyAndClear each transition in tests
Rather than only checking the number of
exiting and entering targets for each transition in
the PostponedTransitionTest tests, always use
verifyAndClearTransition to confirm the epicenters and
exact view instances are correct.

As the epicenter is almost always wrong if the
set of entering/exiting views is wrong, this
also changed the verify ordering so that the more
useful information is checked first.

Test: existing tests still pass
BUG: 147749580
Change-Id: I176b59ea0e6ba32300ffe3cc5d16e678ae1a4dee


GitOrigin-RevId: 79fc88305070516c1329253c2b37c604f0ec44cb