Composable Emits ===
Relnote:"
Prior to this change, the compose compiler plugin would non-trivially intercept calls to constructors inside of a @Composable function if there was an “emit operator” method found on the resolved `composer` in scope. This call resolution strategy was confusing to users and did not follow standard kotlin resolution rules. This resolution existed mostly as a vestigial concept, and is no longer core to the mental model of Compose.

This commit removes all non-standard call resolution that the compose compiler was doing by introducing top-level emit composables that act as a more explicit, albeit verbose, substitute to the emit operators. Although there is additional verbosity to these calls, we feel the benefits of following standard kotlin resolution rules and dramatically reducing the API surface area of the compose compiler plugin more than justifies it.

In order to make this change, a few breaking changes to the runtime and to core needed to happen. They are outlined below:

Compose Runtime Changes:
- `Applier<T>` was changed from a class to an interface
- Added `AbstractApplier<T>` class
- `ApplyAdapter<T>` was removed in favor of `Applier<T>`
- `Composer<N>` was changed from open to final
- `escapeCompose` and `NullCompositionScope` were deprecated as they are no longer necessary.
- `ComposerUpdater<N, T>` was renamed to `Updater<T>`
- a `reconcile` method was added to `Updater<T>`
- existing `compositionFor` overloads were deprecated, new overload was added accepting an `Applier<T>` in lieu of a composer factory.
- two top-level `emit` composables were added

UI Core Changes:
- `UiComposer` was removed in favor of a `UiApplier` implementation
- `VectorComposer` was removed in favor of a `VectorApplier` implementation
- `VectorScope` and all composables that were extension functions to it are now top-level composables and `VectorScope` has been removed
- a top level `emitView` composable was added to make emitting views easier
- `subcomposeInto` no longer requires a `Context` parameter
"
Fixes: 158123804

Change-Id: I5205af707238a70d600c105843cd99e88a5381e0


GitOrigin-RevId: 19285b159b14cbb21d26af97224dbbe6a0b8909a