Add MutableVector, like ArrayList, but doesn't implement List
Relnote: "MutableVector is a new collection that does not
implement any of the standard Collection interface. This
collection offers speed above other requirements and is
intended to only be used in internal implementations."

The performance difference appears to vary between slight improvement
(~15%), with many operations being about twice as fast and removing
elements at the front of the list being 70x as fast.

Bug: 147516363
Test: New performance test compares ArrayList and MutableVector
Test: New test for MutableVector
Change-Id: I8ebd6bf1407595d45eee2a22c1d1db3f38f4ab0b


GitOrigin-RevId: b96a3a54d3d34eefdf34ecd5fd153d7d3af03016