Composition disposal improvements
Dispose of the contents of a composition directly rather than performing
composition with empty content. This allows composition disposal to be
performed more cleanly in response to errors in composition.

Move some composition invariant checks to 'check' vs. 'require' to throw
IllegalStateException rather than IllegalArgumentException.

Ensure posted handler messages that depend on the parent view in
AndroidComposeViewAccessibilityDelegateCompat are removed when the host
view is detached from the window.

Test: existing tests
Relnote: "Applier now requires a clear() method for disposing
compositions"

Change-Id: Ibe697b06ea885852731d029ef56da657b5f290dc


GitOrigin-RevId: 8dec9b77807ec597499539facaa13db6791c9b62