Add rememberCoroutineScope API
Add a mechanism to obtain a persistent CoroutineScope scoped to the
composition at the point of call. These scopes are intended to be used
for launching UI-scoped operations from event handlers, as opposed to
operations that should begin based on presence or absence in the
composition itself.

This API is meant to act as a companion to launchInComposition, and to
some extent a pressure relief valve for the pattern of using onCommit or
launchInComposition to respond to state changes by way of passing those
observable state values as key parameters to these existing APIs.

Relnote: "Added rememberCoroutineScope() to obtain a managed
CoroutineScope in composition for launching jobs in response to events."

Test: SuspendingEffectsTests
Change-Id: I0e264f3f74fc520c6e57051000a513a52f2d07ce


GitOrigin-RevId: 952489cf840c4fcfba2ba57b9b5c6d98828dc464