SubcomposeLayout
It is an analogue of Layout which allows to subcompose the actual content during the measuring stage for example to use the values calculated during the measurement as params for the composition of the children.

Possible use cases:
 * You need to know the constraints passed by the parent during the composition and can't solve your use case with just custom Layout or LayoutModifier. See WithConstraints.
 * You want to use the size of one child during the composition of the second child. Example is using the sizes of the tabs in TabRow as a input in tabs indicator composable
 * You want to compose your items lazily based on the available size. For example you have a list of 100 items and instead of composing all of them you only compose the ones which are currently visible(say 5 of them) and compose next items when the component is scrolled.

The implementation is currently based on an added concept of virtual layout nodes which is considered temporary and will be replaced with a proper subcomposition api which allows to subcompose multiple times into the same LayoutNode when it will be implemented.

Relnote: SubcomposeLayout is added. It is a low level primitive which allows to compose the children during the measuring if we want to use some values available only later during the measure for the subtree composition. For example WithConstraints is not implemented using SubcomposeLayout.
Test: new tests, manually with TabRow, WithConstraints and LazyColumnItems
Change-Id: I25cc8cfe8382db1ef61e93866ba08f4668cbc734


GitOrigin-RevId: 2ca07015c10c37f2136be50066aae55136060888