Update POM URL to point to specific release
Currently, the POM URL points to generic Jetpack or AndroidX homepages. This CL points the POM URL to the specific release in question, for example, https://developer.android.com/jetpack/androidx/releases/fragment#1.3.0-alpha06. This will show up on gmaven and in the version updater plugins as a helpful way to get release notes.

Test: manually ran createArchive and checked the outputted URLs.
Change-Id: Ic314cc6e6c920742129fd52463b81a10f35f620e


GitOrigin-RevId: b50d5005169f75c312e76f5a2da8b7d7887bc81e