Don't try to validate snapshot versions
This CL fixes an issue in playground projects where dependency validator would
try to validate the dependency that would be replaced with the project.

This validation already runs on AOSP with the right version so it is best
to ignore it here. We could've changed this to use resolved dependencies instead
but that runs into a risk when different projects move to beta in sync where the
playground snapshot buildId will be old and fail the build.

Test: ./gradlew :room:room-runtime:verifyDependencyVersions
Change-Id: I04f08d13834eb131365e2e9a20ba3feab75af3f3


GitOrigin-RevId: 27bc4d603109829e3e7450aa04c11641bfabb482