Change default mutation policy of mutableStateOf() to be structural
Fixes: b/159456611
Test: existing tests pass

Relnote: "The default mutation policy of `mutableStateOf()`,
`ambientOf()` and `savedInstanceStateOf()` has changed to be
`structuralEqualityPolicy()` instead of `referentialEqualityPolicy()`.

The default for deciding if a new value assigned to a
`mutableStateOf()` instance is considered a change now defaults to
using `==` instead of using `===`.

See https://kotlinlang.org/docs/reference/equality.html

`ambientOf()` and `savedInstanceStateOf()` use `mutableStateOf()`
in their implementations so they were changed to be consistent with
`mutableStateOf()`.

Using structural equality more closely matches developer
expectations.

For example,

    val state = mutableStateOf(1f)

 followed by,

    state.value = 1f

will no longer be consider a change to `state` and uses of
`state` during composition will no longer need to be recomposed.

This is a breaking change but, in most cases (such when using
classes that do not override `equals()`), this will not have a
noticable effect on an application.

Classes that do override `equals()`, such as `data` classes,
might see a performance degredation as their `equals()` methods
are now, by default, called when assigned to a `mutableStateOf()`.

The previous behavior can be restored by adding the policy
parameter `policy = referentialEqualityPolicy()` to calls to
`mutableStateOf()`, `ambientOf()` and `savedInstanceStateOf()`.
"

Change-Id: Ic21a772051c4f891c655c9bd816ebb360ce19a81


GitOrigin-RevId: ed0866f77a239cd96fb85e1cb3e3ba4fcd781c9a