Update emit to allow skipping some or all of the sets
Inlining `Row` and `Column` caused using them in conjunction with
`composed` modifiers regressed recomposition performance because
the materialization of the modifier would no longer be skipped
(as it was inlined) and modifying any state object read during
materialization (such as a animation) would invalidate the caller
instead of just the emit of the column.

This change allows emits to perform expensive caluculations in a
skippable region that is skippable and restarable independent of
the caller.

Relnote: """Custom emits can now declare that one or more of its setters
can be skipped and recomposed independently of the emit."""

Test: ComposeBenchmarks#benchmark_10_NestedRowColumnsWithModifier
Change-Id: Ibbd13068440252724db405b26d1f6be179e80411


GitOrigin-RevId: 178ada5bcdbb8bcba8b682c49dbd92bb3b89f4f0