Fix the recompose benchmarks
This change add creating  a mutable snapshot during recomposition
benchmarks.

The change to remove weak reference broke the recompose benchmarks
as they, incorrectly, relied on behavior of the old observation
system that allowed it to skip running recomposition in a mutable
snapshot. This also, inadvertently, under reported the cost of
recomposition as it avoided the cost of observation.

Test: CompositionBencmarks
Change-Id: Ibf615b9d7952ba30269cadc008bab851f66f59aa


GitOrigin-RevId: 7b46c0bee3dfd9402cb6ae990c1b3c6906654aa1