Add compose integration for navigation
This change introduces the NavHost Composable which can be used in
conjuction with a NavHostController to manage the in app
navigation for compose. This is built based on androidx Navigation and
inherits much of that infrastructure, but uses extentions to make it
compatitible with the compose framework.

Relnote: "Implementation for compose Navigaiton"
Test: Added NavHostTest
Bug: 161472731
Change-Id: I5c8814743eb99ac4135283a9aa6a81c72f9e7443


GitOrigin-RevId: b7306553b052944cbecfe5b9e1087cc73153139c