Move cursor to CoreTextField
This CL moves cursor to CoreTextField to unblock any TextField developer to gain access to cursor functionality if they desire so. There's an ability to remove cursor by passing Color.Unset as a parameter value (this is also a default, so cursor behaviour for now is opt-in).

Additionally, this work lays a future foundation for cursor drag integration and integration with selection controls in general, namely:
* We will be able to reuse cursor in range selection handles to make then customizable as well
* we will be able to longpress and drag cursor even when no range-selection is active, while maintaining overall proper layering (state will live in CoreTextField)

Bug: 161566455
Relnote: CoreTextField now supports cursor functionality
Test: Corresponding tests changed and added a new one

Change-Id: Id23aa3de7c74074fb9c0d37a42490b6c49f691f3


GitOrigin-RevId: 7b3e381f8fe6164a8e4ef3a1ff5f604353b10cc8