Remove Size operations that consume Offset parameters
Relnote: "Removed operator methods on Size
API that consume Offset parameters. The
result of these operations is unclear
and the type of result is unexpected
for these. Ex. should size - offset return
an offset or a size result with the difference?

Also removed deprecated methods on Size class."

Updated DrawScope usages of operators that
consumed both Size and Offset types
to have explicit calculations done.

Change-Id: Iec902c3664d8669534b8d7eca2b572e6a8d2838a
Fixes: 166514214
Test: Re-ran compose tests


GitOrigin-RevId: b371acdb2499b1f74bdccfd8c8498586ed944d19