Fix crash and items of LazyColumn being incorrectly drawn
If the child LayoutNode was not placed its layers invalidation is still scheduled and we ended up drawing not measured node or just drawing the incorrect state of the node. And then we didn't redraw it correctly next time when the parent finally placed the node as it was already invalidated and not dirty anymore.

Test: InvalidatingNotPlacedChildTest, manually
Relnote: Fix for items of LazyColumn being incorrectly drawn sometimes, this was also causing crashes in some conditions.
Fixes: 163066980
Fixes: 165921895
Change-Id: Ibcf1745e40606f6a38d9eb90f915443935d34403


GitOrigin-RevId: 308929cf794a4e3cb1c1adca86bc62ee32ad1613