Skip version validation for Playground's SNAPSHOT_MARKER
Version validation is currently causing a build failure for jetpad's
buildInfo task which naturally runs as part of bOS.

Dependency substitution seems to happen after this step, so buildInfo is
trying to parse SNAPSHOT_MARKER when checking if that version has
already been released on GMaven.

This change skips version validation for SNAPSHOT_MARKER, since it by
definition is an unreleased target.

Test: cd paging && ./gradlew bOS
Change-Id: I417acbabed6be26dbb70d4c19ddeee1352250907


GitOrigin-RevId: f2d43dd9d6bc29112f2cb3911ba5db405d9296e4