Keyboard hidden when moving between textboxes
We call inputMethodManager.hideSoftInputFromWindow() when we lose
focus and inputMethodManager.showSoftInput() when the next text
field gains focus. Calling show() after a hide() causes the
keyboard to flicker and end up hidden.

This fix sends the show/hide messages to a Conflated channel so that
only the latest message gets delivered.  Then we either show or hide as
needed.

Fixes: 166197099
Fixes: 169721891
Bug: 164859446

Test: ./gradlew :compose:foundation:foundation:connectedCheck -Pandroid.testInstrumentationRunnerArguments.class=androidx.compose.foundation.text.CoreTextFieldSoftKeyboardTest
Change-Id: I6b171f80b30c9b355a3bae465b5821139f9a264d


GitOrigin-RevId: 276fd5b474750233997f863092d57be6883764ba