Safer scroll position handling in LazyColumn/Row
Introduced private FirstVisibleItemHolder to hide the whole fragile logic where we have to duplicate source of truth and keep firstVisibleItemIndex and firstVisibleScrollOffset in both State and regular variables. This abstraction makes it impossible to modify one variable and forget to modifier another one or to apply different values by accident.
Plus covered this logic with few more tests.

Test: LazyColumnForTest, LazyRowForTest
Change-Id: Ie6d19025e2a3f8558acea5653f23e3720b4c2a21


GitOrigin-RevId: 2a763c44dcb3cf9a620af39e3192bdfce4cd23c1