Add FLAG_SECURE on Dialog and Popup.
Added a new way on how to pass platform specific parameters for popups
and dialogs.

Also renamed children to content. And made expect / actual to be
internal for dialog as it is for popup.

Bug: 143778148
Bug: 143778149
Test: Added tests for both dialog and popup.
Relnote: "Popups and dialogs now inherit FLAG_SECURE from parent Window.
Also added option to configure this explicitly".

Change-Id: I649663482e91757df751315b03fee9867b580e96


GitOrigin-RevId: a9b67eca31e73a3f975b53e22a9e5983e2d76864