Fix for crash in SubcomposeLayout when the view was detached
We didn't check that the layoutnode is attached before doing subcomposition which can cause a crash if the recomposition is scheduled, but the View was detached. We need to skip the recomposition. Once the View will be reattached all the LayoutNodes will request remeasure and it will cause recomposition to happen.

Fixes: 169944654
Test: covered the issue with a new test in SubcomposeLayoutTest
Change-Id: Ib62107c777218b101c27ec04d04f02c25c791fc1


GitOrigin-RevId: 951cd444c0f5d3046356c3e0c0d6d24ef4b7aca6