use reflection to access incubating API.
project is using AGP incubating APIs that have changed between
alpha06 and beta1. Androidx is using alpha06 and cannot upgrade
yet, however PSQ runs with the latest -dev version.

Fixed by removing all type accesses to the the API and replaced
it with reflection based access. It's not pretty but it allows
to build successfully with both alpha06 and beta1 using their
respective API.

Once androidx can move to beta1 or above, this should be relaced
with proper new API access (an example is provided as a comment
in this CL).

Test: Existing, can build with -dev and alpha06
Bug: N/A
Change-Id: I107c74448d861ef40cd95118c524a97baaffd0ad


GitOrigin-RevId: 1759d919d35f691dbb9680d8c5a977155b876124