Fix playground projects.
This CL fixes two issues in playground which broke after the main
project is refactored for compose.

It introduced a required environment variable to pick the list of
projects. For playground, it doesn't make sense since there is only
1 set of projects declared in gradle so I removed that option from
the playground studio launcher.

The change also broke the playground's include functionality which
parses the main settings file. Now, there are more arguments to
includeProject which made it skip the regex. I've updated the regex
to allow more parameters after the first two (project path and directory).

Bug: 171017446
Test: locally verified with paging & room
Change-Id: Ib41126ce7a3aab572c016420c1375599512ab258


GitOrigin-RevId: 7bdac5188136a0756e598e3c60b3881a9cf60aac