Forced Parameter for FocusManager.clearFocus()
Adding a forced parameter to clear focus that allows a developer to
clear focus even if the focus is in a Captured state.

Right now if developers use a component that internally has multiple
focus nodes and one of them captures focus, they will not be able to
clear focus. This API makes it possible to clear focus.

With this API we can simplify some of the logic in requestFocus().
When a focusedComponent is detached, we can forcefully clear focus
and so we no longer have to worry about the order in which the
modifiers were detached.

Also fixed a bug where we didn't update focusedChild when an
ActiveParent was detached. Added a check that throws an error
if a modifier is in the ActiveParent state but does not have
a focused child.

Bug: 163150504
Relnote: Allow a developer to forcefully clear focus.
Test: /gradlew compose:ui:ui:connectedCheck -Pandroid.testInstrumentationRunnerArguments.class=androidx.compose.ui.focus.FocusModifierAttachDetachTest
Change-Id: I9037255fff69d59e9596632c9e63875fdbd9df89


GitOrigin-RevId: d0dfc6ff66976340cc12a420935a171e6eb1aa25