RestorableStateHolder
RestorableStateHolder allows to save the state defined with [savedInstanceState] and [rememberSavedInstanceState] for the subtree before disposing it to make it possible to compose it back next time with the restored state. It will be used by the Navigation compose library and is flexible enough to be used by any other similar use case or third party implementation of the navigation pattern. To demonstrate it I integrated it with our existing Crossfade animation component in CrossfadeDemo. This mechanism will also power the saving of the state for items which are no longer visible in LazyColumn to be able to restore item's state like inner LazyRow scroll position when the item with LazyRow would became visible again.

Relnote: New experimental api RestorableStateHolder. It allows to save the state defined with [savedInstanceState] and [rememberSavedInstanceState] for the subtree before disposing it to make it possible to compose it back next time with the restored state.
Bug: 166586419
Test: new tests in RestorableStateHolderTest
Change-Id: I66884b1e65f716de7936c4fe9e573efc6539b80f


GitOrigin-RevId: c70e44710c9295c6a9610f92856ac96e84cf90fe