Fix issue in the slot table that caused anchors to be inconsistent
When the slot table is exactly full (capacity == size) and the
gap is moved, the anchors are not updated to reflect the new gap
position. When an insert, delete or move is then performed the
anchors will start tracking the wrong groups.

Now unit tests check for this by validating that recompose scopes'
anchors now point to the groups that contain them after each
recomposition. This caught the error in existing unit tests so no
new test was needed.

Fixes: b/171689853
Test: CompositionTests
Change-Id: I6b660757730696e037afafac6e1b9e6446c2c1d3


GitOrigin-RevId: ff31d82d16a87d803eb51eab20404d45116c1ab5