EditText in AndroidView crashes when using hardware keyboard in emulator
AndroidCompose tries to route hardware keys based on the currently focused focusModifier. However the embedded view does not add a focus modifier, and so this crashes with the following exception:
java.lang.IllegalStateException: KeyEvent can't be processed because this key input node is not active.

This CL fixes this issue. Also adds a demo app and unit tests.

Bug: 165034731
Fixes: 165034731
Test: ./gradlew compose:ui:ui:connectedCheck -Pandroid.testInstrumentationRunnerArguments.class=androidx.compose.ui.viewinterop.EditTextInteropTest
Change-Id: I25fb88f8a6cdda52a69f0ab2a4482569015503b1


GitOrigin-RevId: 68c70e3a04007ca09c6953d81bd3aeb70406339b