Wait until DialogFragments are attached
Instead of assuming that all DialogFragments are
already attached when onRestoreState() is run
(which is only case when using Fragment 1.3+ due to
the reordering done in b/171605525), wait
for each DialogFragment to be attached if it
doesn't exist in the FragmentManager already.

Unfortunately, the API we would want to use here,
FragmentOnAttachListener, is only available in
Fragment 1.3, so this fix requires the involvement
of NavHostFragment to manually pass the
onAttachFragment() callbacks to the
DialogFragmentNavigator.

Test: tested in integration test app (as that still
relies on Fragment 1.2.5)
Fixes: 172654190

Change-Id: Idbd849139d4c4656e38c9c98f47075e00f168c71
(cherry picked from commit 3711505bb840c02aa33d06ab2fa4c757150f1278)