Handle semantics bounds change and throttle window content change events
When semantics node bounds change, accessibility needs to send out
a subtree change event to update the accessibility cache and rerender
accessibility focus. Note that bounds change of a LayoutNode without
semantics will also affect its children which might have semantic, so
every LayoutNode in the subtree is processed. The final window content
change event sent out is guarenteed to have a source node which is not
merged to a parent (because accessibility doesn't recognize nodes merged
to parents).

This CL sends these events in throttled batches every 100ms, in order to
avoid degrading performance particularly in animations.

Bug: b/153643868
Relnote: N/A
Test: Two new tests in AndroidAccessibilityTest

Change-Id: Ie6541e0ce7c757da78cb2ae41cdcc1d3abb3aeb6


GitOrigin-RevId: 694db0af0dbe50c4757410b5d29f46ae63a18389