Add a new docs-snippets project to track changes in DAC docs
The official documentation (like
https://developer.android.com/jetpack/compose/themes) has a large number
of code snippets that are hard-coded. Maintenance of these articles is
proving very hard given that there are breaking changes every 2 weeks
and that they require subject matter experts.

With the snippets in AOSP, we can track changes to these files and
easily detect the pieces of code that need updating.

If we put this on a separate project, depending on artifacts, we
wouldn't benefit from automatic refactorings, which is what most of the
breaking changes are at this point.

Bug: 173574611
Test: N/A (build = ship)
Change-Id: Id4a0ba4a3bfcf183aeb8a6bf0d9ca035b48fdb3a


GitOrigin-RevId: ca429dd718b811decb3fe4570637b22d63e32733