Add option to drive clock by MonotonicFrameClock in tests
It is now possible to let the test clock be driven by the
MonotonicFrameClock instead of the Choreographer. This comes at a cost
though: time control is not implemented in this variant. That is OK,
because this is only meant as a temporary solution to investigate moving
all clocks to the MonotonicFrameClock. It will be removed once we moved
over.

Fixes: 173402197
Test: MonotonicFrameClockTestRuleTest
Relnote: "Temporarily added option to let the TestAnimationClock be
driven by the MonotonicFrameClock"

Change-Id: I1403ba3d82adc530d885192aa696c4363456a4c1


GitOrigin-RevId: 33c9aa2d32c2f39d0e7a3934cf14f30b5d0c8513