Add maxLinesHeight modifier to fix scrolling issue when maxLines is not infinite.
For scroller to work, CoreTextField needs to report the real height the text is occupying. Currently when we TextFieldDelegate lays out the text, it coerces its height so that the reported height is no more than the maxLines parameter require. So I moved the maxLines constraining logic to the BasicTextField using the modifier.

Test: new tests in MaxLinesHeightModifierText and TextFieldScrollTest
Fixes: 173296792
Relnote: "Removed 'maxLines' parameter from the CoreTextField. If you'd like to constraint the height of the text field by number of lines, use BasicTextField instead."
Change-Id: Iec0026dfe939c60e110304db6cda0fdaa57a034f


GitOrigin-RevId: dc066230c77406f5b1d7d63550939b1b6b963013