Update KSP dependency to 20201204
This version is important because it also moves KSP's
kotlin dependency to 1.4.20.

I had to specify more versions for testing because we were
getting kotlin compiler dependencies from kotlin compile
testing, which is still on 1.4.10. Now compiler-processing-testing
artifact explicitly specifies those version to match whatever
AndroidX is using.

I've also removed workarounds for the following issues
as they are fixed.

https://github.com/google/ksp/issues/170
https://github.com/google/ksp/issues/174

Bug: 160322705
Test: existing tests
Change-Id: If948bcb762d1fad4911cf833a85cda6a36e7c787


GitOrigin-RevId: c74f6d8bb25b24bfe78cef9cb4e080ef988d483a