Fix direct transition from VIEW_CREATED to CREATED
When a fragment is replaced before the FragmentManager
has reached the ACTIVITY_CREATED state, it can transition
directly from the VIEW_CREATED state (i.e., having
`onViewCreated()` called on it) to the CREATED state which
should trigger `onDestroyView()` and the removal of
the fragment's view from the view hierarchy.

By ensuring that the createView() and destroyFragmentView()
calls are symmetric in moveToExpectedState(), we can
avoid cases where onDestroyView() is skipped and avoid
cases where the previous fragment's view is still in the
view hierarchy.

Additional logic has been added to StrictFragment and
StrictViewFragment to take into account this transition
edge.

Test: new FragmentReplaceTest passes
BUG: 168987859
Relnote: "Fixed an issue where the fragment's view would
remain in the view hierarchy if it was replaced during
its `onCreate()` or `onViewCreated()` methods."

Change-Id: I8a7d5aed2c5648d3b898bb1cc1469f90cf5f1abd


GitOrigin-RevId: 53c68e37445eabb266b2ee48e507e5301847ff73