add Semantics role API
To identify widgets, we have traits and roles. From the discussion we
think using roles is the preferred approach. Also, I'm not sure we
still want the string format roleDescription API in this case, so it is
currently not added.

Relnote: "Add Semantics role API and add Role as a paramter to
clickable, selectable and toggleable SemanticsModifier. Changed
Modifier.progressSemantics so that Slider can also use it."

Test: tested with demo app with talkback. Unit tests added.

Fix: b/172092161, b/153637146

Change-Id: I216cd5b9118581e569c48a095f009c841ed4418b


GitOrigin-RevId: 6142074942b7ef972a5f76ea2954631f406ff633