Make SemanticsNode tree zSorted and use it for a11y touch explore
For accessibility, when touch explore is enabled, touch an element will
dispatch a hover MotionEvent. The element will need to send out hover
accessibility event to notify accessibility services.

In compose, for touch events handling, any touch event filter will block
all touches to its later sibilings on the touch point
(but can pass touches to its parents),
due to the hit test before we dispatch touches. And in principle, any
touch event filter on a layoutnode should be associated with some
semantics. So when we want a node to block its siblings (for example,
drawer, bottom sheet, backdrop), we will have a touch event filter on
this node because the blocked contents can be anything, and we must block
all touches to the blocked contents. And when we have a touch event
filter, we will get a semantics node reported so we get the blocking
node.

The transparency for blocking nodes is not an issue in compose because a
transparent node without semantics will not generate a semantics node. A
transparent node with some touch filters will block interaction with its
siblings drawn below it, so a transparent node with touch filters drawn
on top of interactable sibilings is not possible (in regular android,
this can be done).

So the zSorted semantics tree is a good proximation for what elements
are touch explorable (we should not aim for perfect here, the android
hover event for touch explore has bugs too, like if an opaque text view
is drawn on top of a switch, touch the textview will place accessibility
focus on the covered-switch instead of the textview. From reading the
code, I think the same will happen for two overlapped opaque textview,
i.e, the accessibility focus will be placed on the covered textview
instead of the top text view).

One drawback we can think of is some blank opaque node is drawn on top
of some non-touchable nodes (like text) to cover the text, we won't get
the covering blank node and we will touch explore the covered text nodes
instead. Note this is a very specific case as it requires the covered
nodes to be non-touchable (if we don't know the covered content, we have
to add some touch filter to the covering nodes). In this case, we can
ask developers to put a blank semantics block on the covering node to
capture the semantics that this node is "covering things". Hopefully,
this should not be frequent. And overall, it is still a big improvement
compare to regular android counterparts.

Fix: b/168732207

Test: Semantics test and accessibility test added.

Change-Id: I500e3d318448a42af8b1405e5f707f9275b8335e


GitOrigin-RevId: 0b82f3306bb844172334a26e9005409acf577c64