Add option to hide system nodes from layout inspector
For now: hard code the packages considered to be system packages.

Adapt the code for stitching trees together because a tree may end
up containing no user defined nodes but should still be able to
contain another sub tree.
Do that by creating a temportary empty node, that is removed before
the resulting tree is returned.

Bug: 136022822
Test: Added unit tests where only non system nodes are expected.
Change-Id: I875e4b898d7cc3a84e3beb99693b0b5321eaf58e
Relnote: N/A


GitOrigin-RevId: 9e35be5d91ce2c33352217ece3c63a246fd3f6d5