Desktop Paragraph tests for Linux
After a discussion with Julia Lavrova (author of SkParagraph) about cross-platform tests of Skia Paragraph, we figured out that it's practically very hard.

So, at least for now, we are making exists Desktop Paragraph tests Linux specific and start writing new tests for the Linux platform

In this test, we assert the paragraph sizes to the ones which are expected on Linux. The values will be different on Mac, we will handle it properly later

Test: ./gradlew :compose:ui:ui-text:desktopTest
Change-Id: I83c551645269aa934e4030ee32d8b8f05bf9cb42


GitOrigin-RevId: 68028cc99733e92fcf78ddfaaade3e8b36e8f5c2