Relax clock locking requirements, and start using scaling_setspeed when possible
Fixes: 174791675
Test: ./gradlew benchmark:benchmark-common:cC
Test: Bramble locked with platform lock script now registers as locked
Test: Ran ShellTest on API 18,21,25,29, some of which were emulators
Test: ./gradlew benchmark:benchmark-junit:installDebug && adb shell am instrument -w -r --abi arm64-v8a  -e debug false -e newRunListenerMode true -e size small -e log false -e notAnnotation androidx.test.filters.FlakyTest -e timeout_msec 300000 androidx.benchmark.junit4.test/androidx.benchmark.junit4.AndroidBenchmarkRunner

Big/little/gold/medium/platinum/bronze core configurations make it
unreasonable to expect all cores to be locked to the same
value. Bramble is a good example of this, with 2 powerful but unequal
cores, and several small ones. It doesn't make sense to only enable
the biggest core, since that destroys the ability for the system to
schedule background work, or even handle trivial parallelism.

Now, we just require that all cores of the same configuration are
locked to the same value.

Additionally, use scaling_setspeed over scaling_min_freq, as some
locking configurations won't be reflected in scaling_min_freq. This
means we have to start using the shell in microbenchmark (but without
the UiAutomator dependency), as setspeed can't be read directly by the
app.

This change also adds protection against the new UiAutomation logic
being initialized on the UI thread. If UiAutomator attempts to connect
from the main thread, the connection fails, which can happen in
microbenchmarks if e.g. CpuInfo.locked happens to be first queried by
the UI thread. See b/193064052 - last `Test` line above shows
invocation that happened to provoke this connection failure in CI.

Change-Id: I23f501f9a54736551ae0b1491bf0dd08d1661bac


GitOrigin-RevId: 38398bc9648f36e7067c11a9034a813a0f6a8299