Handle metadata publishing in external artifact checker
This CL fixes a bug in CheckExternalDependencyLicensesTask where it
didn't handle artifacts that are published with gradle metadata.

More specifically, when a artifact has gradle metadata, the final file
might be copied to gradle caches with its module name added to the
artifact. To handle those cases, if the artifact is not in prebuilts, we
try to find the artifact from local maven repositories.

I've kept the `prebuiltsRoot` property for license checks instead of
replacing it with the declared maven repositories just in case someone
in the future adds a repository outside prebuilts.

Bug: 177366060
Test: ./gradlew checkExternalLicenses

Change-Id: I81cd1956faa713c5455561f8d52020dbd20609c6


GitOrigin-RevId: fe98af73c536fe04162b85108bd4fe58320032f3