Fix ComposeView Activity leak
Cancel the watchdog that unsets the window recomposer reference from the
root view of a window when the recomposer shuts down if the root view
becomes detached from the window. This was keeping the host Activity
reference alive indefinitely in the presence of a global recomposer that
never shuts down.

Fixes: 177107788
Test: WindowRecomposerTest.activityGarbageCollected
Change-Id: Iee12af1d1dfb781846ab7397e7a1b69916e1bd83


GitOrigin-RevId: 6626dee980e3bf4b5a79a8d3df7ada4ae78d8b1c