Fix drawer closing on click and blocking the rest of the ui.
SpringSpec is settling slow and was incorrect anyway, so I dug into the DrawerLayout.java implementation and took a constant from there to provide a better UX when drawer is closing. Filed b/177571613 to fix this properly.

Also added stub click listener to block scrin and UI below drawer sheet to be clickable when drawer is open

Fixes: 169970157
Fixes: 171601051
Test: added new
Change-Id: I997bfbe3dea96f0325617faf1e002591d348025b


GitOrigin-RevId: ecd05c42e2e9e2f9e91f4b2386194bfabaa64876