Add PaneTitle Semantics API and apply it to drawer
When the drawer opens, we use Pane API to announce the
drawer title. Pane node appear/disappear/in screen/out of screen are
handled in accessibility delegate. Added Strings.kt in material
to hold string constants.

Note this implementation is based on pruned accessibility tree
(aosp/1540344), if we later decide not to prune accessibility tree
later, this implemention needs to revised. Here is a reference to
an implementation that doesn't depend on pruned tree: aosp/1482398.

Relnote: "Add SemanticsProperties.PaneTitle API."

Test: tested manually with talkback. When the drawer opens, it says
"Navigation menu". When the drawer closes, it says "Jetpack compose
demos". Integration test added in AndroidAccessibilityTest.kt.

Bugs: b/153643868
Change-Id: I20d5a27adec172993b792a4cfa7df37280aa0425


GitOrigin-RevId: 96e8156091159ee6c2702478d6f9425daf72954e