Explicitly uses rememberRipple() in Button and FloatingActionButton
To customize the ripple used by Material components, RippleTheme should be used - providing a different indication via AmbientIndication is not the correct behavior here, as it will only customize Material components that do not explicitly create ripples, in cases where they don't need to customize the radius / bounded behavior. With this change there are no other components in compose:material:material that do not explicitly create a ripple.

Test: n/a
Relnote: "The ripple used in Button and FloatingActionButton can no longer be customized by providing a new Indication through AmbientIndication - this was never intended to be a way to customize these components and this now makes these components consitent with other Material components. To customize ripples across an application, see RippleTheme."
Change-Id: I546c59bf53904435e14590ed4b3ca9973704bff3


GitOrigin-RevId: c918a7e2e9f10ac4948c6b32fd040176073234a2