Add isInfinite to Animation and VectorizedAnimationSpec
Will be used to facilitate special handling of infinte animations in
tests, so the test framework won't await these animations indefinitely.

Bug: 151940543
Test: IsInfiniteTest.kt
Relnote: "Added `Animation.isInfinite` and
`VectorizedAnimationSpec.isInfinite` that signal if an animation is
infinite or not. This can be used in implementations of animations to
have special handling of such animations. For example, a special \"time
remaining\" message can be shown, or during tests the animation can be
cancelled to prevent waiting for idleness indefinitely."

Change-Id: Iebb05e9d158b4fe81d037ab28e113da4926c50cd


GitOrigin-RevId: f0649b27c6df6fa151202e3f23913998805fb398