Removes `toggle` parameter from DropdownMenu
This change makes DropdownMenu behave similar to a Popup, and reduces the complexity of the API.

Fixes: b/173610319
Fixes: b/172096706
Fixes: b/178817668
Test: MenuTest / demo
Relnote: "Removed `toggle` and `toggleModifier` parameter from DropdownMenu, and renamed `dropdownModifier`, `dropdownOffset` and `dropdownContent` to `modifier`, `offset` and `content` respectively. DropdownMenu now behaves consistently to `Popup`, where the parent layout is used for the position of the menu. In most case you can move `toggle` to be a sibling of `DropdownMenu` and wrap both in a `Box`. See the updated sample in the documentation for more information on usage of this API."
Change-Id: I884fbf5ca080e3afea00146c5a7a869247b3613e


GitOrigin-RevId: dc48bf4831fac8e21d8a13121970accc1f95a715