Output Studio v2display, with trace links
Fixes:181593679
Test: Studio output with v2display support, and test app
Test: ./gradlew bench:b-c:cC bench:b-j:cC bench:b-m:cC bench:b-m-j:cC
Relnote: N/A

Outputs links to traces, both for each metric, and as a list at the
end of the benchmark.

To support old versions of Studio, we still output the existing
`android.studio.display.benchmark` format. The new instrumentation
result string is labeled with "v2display" to prevent older versions of
studio to attempt displaying it, as they won't be able to properly
display its links.

Also fixed Stats.stdDev to report 0 for size=1 stats objects, and
equals/hashcode, as these were affecting tests.

Change-Id: I81a19c2122e53a80cc3fec6f00a8430a32dd2896


GitOrigin-RevId: 559b9d0caa2431bbf3f9088f485b3a227d5b7376