Introduce ExperimentalProcessingAPI annotation for room processing APIs
These APIs are specific to room but there has been an external request
to make them usable so people can try KSP easier.

We have no intentions at this point to support these as public stable
APIs but using this annotation, we can let people use them and
understand the risks. Eventually, this artifact can grow into a real API
so getting more people to play with it is actually a good option to test
it.

In theory, we should be adding this API to every single public method in
compiler-processing. Even though it is do-able, it is a lot of
unnecessary and confusing code.

Instead, only the entry point (XProcessingEnv) has the annotation since
this library is not usable w/o that entry point. Room-compiler opts into
it via its gradle file to avoid the churn there. The testing module, on
the other hand, inherits the annotation to ensure developer won't try to
use them without opting in.

Also enabled publishing for the testing artifact

Bug: 160322705
Test: Run tests with -Pandroidx.allWarningsAsErrors
Change-Id: Ib165b501fa57e2c69a5e558667d6770f0d63cc55


GitOrigin-RevId: 186c42299f239aabfc9c56d0533113706d92fd03