Desktop window onKeyEvent & onPreviewKeyEvent callbacks
It adds the ability to setup onKeyEvent and onPreviewKeyEvent at the very top of KeyInputModifiers hierarchy.

It removed Modifier.shortcuts as badly designed, not flexible enough, and in general incorrect feature: it's impossible to say at that level what keys are pressed now in all cases. Window can just lose focus and RELEASED events are not to be delivered to the modifier. Instead of Modifier.shortucts the raw KeyEvents can be used since they now allow access to pressed modifiers. Or some higher-level abstractions can be built on top of KeyEvents in a user code.

Now handling of KeyEvents goes through a single hierarchy of KeyInputModifiers, making API uniform and implementation is more straightforward.

Test: manual
Change-Id: I237112f340362acc0aff28cdc1d41df6510b5401