Adds an internal lint check to warn if creating an unnecessary Iterator
When iterating over a List (or any index based collection) it is cheaper to loop over indices and get each item instead of allocating an Iterator object - in hot code paths such as in text / runtime, this can make a lot of difference.

ListIteratorDetector warns if:

- Using `for (a in b)`, where `b` is a List
- Calling a method that is an extension function on Iterable, on an instance that is a `List` (as the only public API of Iterable is `iterator()`, which creates an `Iterator`

This CL also migrates existing usages away from allocating functions, and adds internal non-allocating utils where needed, as we can't add new APIs to `ui-util` during beta.

Note: these optimizations are only important for low-level internals, as code relating to text / layout / runtime will be invoked across a large amount of a composition. For user facing code this sort of optimization is not very meaningful, and external developers should not worry about the impact of allocating an Iterator in their code.

Fixes: b/181883246
Fixes: b/181884656
Fixes: b/181968452
Test: lintDebug
Relnote: N/A

Change-Id: Ia7146ed61bf73b8b20457fce03ce18e0f17f4d81


GitOrigin-RevId: c6e92aaf4365432755eeb78ea71873ceaa46d027