Make Modifier.scrollable use Modifier.draggable.
Refactoring CL to increase code reuse and make sure that the Modifier.draggable fits the overall scoped states idea.

Unfortunatelly, to make this work properly with mouses for now one more parameter was needed, so I added an internal overload. Even with the overload, there's still value in making scrollable to use draggable, as is simplifies the logic.

Bug: 180030424
Test: DraggableTest, ScrollableTest
Change-Id: Ia0262a253ca0b6c88253806aada719c39ee9b497


GitOrigin-RevId: 3c66c713f2892672e10a3d9523c767a04dba96e4