Return exactly maxElements from List/Arrays
For Iterables we were returning the looking at the first 5 elements
(including null values) and returning just the non null values.
After this change we return exactly 5 actual non null values.

Also: abandon -1 as index value for indexes that match the
elementIndex. This just caused trouble on the client side.

Also: add special case for Offset, to be able to call expand
and get the x & y values from their offset. This was caught by
the test.

Bug: 180952854
Test: Unit test for longer String List with null elements added
Change-Id: I892cad4284ba601160c1599c3e183a13b8d4d895


GitOrigin-RevId: aa14b8bfa5344267e9e8ac3484a2a08cf81b431e