Rework animateAsState to support frame-by-frame interruption
`withFrameNanos` doesn't resume until the frame is complete,
as a result, if animation values are not updated in the
`withFrameNanos` block, it'll wait til the next frame. This
becomes especially problematic when the animation job gets
cancelled before the next frame comes.

In the case of frame-by-frame interruption, such as setting
`animate*AsState` target based on dragging, the animation job
gets canceled every frame, and the value never gets udpated.

This CL uses a channel to store the incoming target values.
This way the animation doesn't get externally cancelled, but
rather handles the interruption with its iterruption logic as
it receives new targets from the channel.

BUG: 182388979
Test: new test added, existing tests on animate*AsState all pass
Change-Id: Ia99e605cf07dd672c7611a976293713045e3ae86


GitOrigin-RevId: 62cc9183d21095e78ebb3a3bc78daa3800f7d3b6