Fix blinking in fade in
The blink is due to the snapTo behavior change when
it became a suspend fun. What was expected was in
the beginning of the fade in, alpha gets snapped
to `intialAlpha`. But the job for `snapTo` ended up
getting executed after the frame. So for one frame,
alphaAnim.value = 1f (what it was intialized with).

Also changed Crossfade Modifier.alpha to
Modifier.graphicsLayer { alpha = ... } to skip
recomposition during fade by delaying the state
reads to draw stage.

Bug: 181156422
Test: New test added

Change-Id: I2b1db76541477f5eda713f420c18b2d0cf09a42b


GitOrigin-RevId: cbe05dd21692e216f00ce8d9c743fc9e038af0b3