Remove extra recomposition/remeasure from CoreTextField
TextFieldScrollerPosition.maximum is read from composition, then changed during measuring which immediately schedules recomposition. In general it would be great to not need recomposition at all for scrollable() but there is no such api yet. But what we can do is to at least have a reasonable default. In most cases we don't expect TextField to be scrollable so we set 0 as a default value so when we change from 0 to 0 the recomposition will not be scheduled

Test: manually, there is no good test infrastructure to cover such fix with a unit test
Change-Id: Ia7b36655f608aefb8cf489a77a24b14b50d35fdd


GitOrigin-RevId: 79c9dce72f3bad07845259502b7121d8b10ce7aa