Rerun LazyColumn/Row content block before scrolling if it is dirty
It is possible that after changing the total count of items of LazyColumn(for example via Paging) the next scrolling will happen before the scheduled recomposition during which we were previously rerunning the items builder block. This scrolling can be triggered via a fling animation of touch input. In this case previously the scroll could try to compose an item which within the bounds before the changing the total count but it is not anymore. This was causing ArrayIndexOutOfBounds exceptions.
With this refactoring we can now rerun the items builder block if it is dirty both during the composition and measuring(done as part of the scrolling) which guarantees we always have the latest items count and item content lambdas.

Test: new tests in LazyColumnTest and LazyRowTest
Fixes: 182029524
Change-Id: I61265c0ee3618538c60bab6b409aedc303e85ad6


GitOrigin-RevId: ba77122567ff5d31f9f68c7125c9699f7a354c84