Implement view interop in accessibility
Layout node with an android view will have an empty semantics
node. We use androidViewsHandler to keep track of mapping from
layout node to view. The android view has an accessibility
delegate to find the parent virtual node.

For touch explore hover event handling, we always
dispatch the hover_enter/move to the views first. This is because
the android view children could be view groups, so we must dispatch
the events to the views and let the views handle it instead of
sending out accessibility hover events ourselves. Android ViewGroup.java
will take care of synthesizing hover enter/exit actions from hover moves.
We dispatch it to the views first and then update hovered virtual view
so that if the hovered view is on top of an android view, the final
accessibility hover enter event will be sent from the virtual view.
If the android view is above the virtual node, the hovered virtual node will
be null and no accessibility hover enter from the virtual node will
be sent. (the final hover enter accessibility event will decide what
node talkback will focus on).

Added method
"public fun findViewByAccessibilityIdTraversal(accessibilityId: Int): View?"
which is a hidden method in android as a workaround to old android versions.

Test: tested manully with demo app and talkback. Integration test added.

Bug: b/180583781

Change-Id: Ie4da95cbe47b54d443d729837b49ee3831759259


GitOrigin-RevId: f3b4f4f28cf61d4434af59d9247a9d2f6d5d70e7