Add OnBackPressedDispatcher observer before NavController observer
If you add a OnBackPressed callback with the NavBackStackEntry as the
lifecycle owner, if the Activity Lifecycle is STOPPED and STARTED again, the navController callback and NavBackStackEntry callback are readded to the Activity Lifecycle in the wrong order. This means that the new callback will not properly intercept onBackPressed() calls from the system, they will instead be intercepted by NavController.

If an OnBackPressedDispatcher is set, we should remove and readd the
NavController observer after we add the callback to the dispatcher.

RelNote: "OnBackPressedCallbacks added to a dispatcher with the
NavBackStackEntry as the LifecycleOwner now properly intercept back
after the Activity Lifecycle is STOPPED, then STARTED"
Test: NavControllerTest
Bug: 182284739

Change-Id: Iff94f9efa9739fa49166ddb2dc11b50d2a4c4a8c
(cherry picked from commit a95121cbd1e87689b3660062acf0fb6d5a04883b)