Allow navigating by routes
This adds support for navigating to destinations via a route. NavOptions
are all added via the NavOptionsBuilder attached to navigate(). You
cannot navigate to NavActions using routes. It is possible to have a
route and a custom destination id, and use either to navigate to your
destination.

This change also includes support for routes in TestNavHostController,
and additional util methods in TestNavigatorDestinationBuilder.

RelNote: "Navigation now supports navigating via a custom destination
route."
Test: NavControllerRouteTest, NavDeepLinkBuilderTest,
TestNavigatorDestinationBuilderTest

Change-Id: Ie1af7f2672acff48dc26869d8ef0c601559f4d43


GitOrigin-RevId: 34f74de9a1783c49dbe116106683af001d1a85d4