Append UI state to macrobenchmark traces
Test: ./gradlew benchmark:b-m:cC
Test: TrivialStartupBenchmark, open traces in Studio build 7302168
Relnote: N/A

We use a manually minified version of Perfetto's
perfetto_trace.proto. This is done to minimize library method count /
binary size, and because with Wire, the full perfetto_trace.proto
creates too-large methods for the Kotlin compiler.

Would prefer to instead use wire's include syntax to handle
minification (below), but this fails to resolve the types of optional
fields within perfetto.protos.TracePacket that we don't need.

    include ['perfetto.protos.UiState', ...]

Note - only tested integration with Studio using
UiState.HighlightProcess.pid on Studio build 7302168, as cmdline
support had not yet landed.

This change was initially attempted with google proto-javalite, which
requires shading the runtime dependency. With that codegen, the full
proto file (not the minified one in this CL) also hit multidex, and
couldn't be indexed by Studio intellisense (even with
`idea.max.intellisense.filesize=99999`).

Change-Id: Iffbf9c470587d05a0640645e0b54c6ce71d0e98f


GitOrigin-RevId: f7ea263a420020e5873c9e153f22599e47241275