Fix accessibility text traverse/selection event ordering
In talkback, when we focus on the edit text and
press volume up/down key, we move the cursor left/right. And if
the verbosity setting is high, talkback announces the phonetic
word. In current talkback version, this relies on the traversed
event received after the text selection change event. So in
Compose, when we perform a traverse action, we record the
traverse event. When we send a text selection change event,
we check whether we have pending traverse event and see whether
the pending traverse event is new enough (in error state, we
don't want a stale pending traverse event). If there is a pending
traverse event and it is new enough, we send the traverse event.

This applies to static text traversal too.

Test: tested manually with compose demo app on basic input fields
and static text on android S and O. Integration tests added in
AndroidAccessibilityTest.kt. Note that I enabled the test. I
run the original test 100 times on pixel 2, api 26 and emulator
pixel, api 26, but I couldn't reproduce the error in b/178524529.
I put the performing action part on UI thread. Will leave to
presubmit to verify.

Fix: b/178524529, b/173153373

Change-Id: I1efebdb2cd806e55e2fa4135cd665f0a7add5842


GitOrigin-RevId: 7df10ca3a307f17f3b541dd37252395b5da08c60