Remove InputEventCallback from public API
The interface is not being used in any public
APIs. It is not possible to create an instance
and pass to a public API.

Initially it was used by hidden Android specific
classes, and now those classes use InputEventCallback2
since the interface does not have to be public.

RelNote: "InputEventCallback interface is deprecated.
It was not possible to use the interface in any
public API; and there was no usage of it in the
code."

Test: ./gradlew compose:ui:ui:test
Test: ./gradlew compose:ui:ui:connectedAndroidTest
Test: ./gradlew compose:ui:ui-text:test
Test: ./gradlew compose:ui:ui-text:connectedAndroidTest

Bug: 184003208
Change-Id: I34a0208fac88f99da33ce9c700a803454dfb27c7


GitOrigin-RevId: 2a814d0c81e2cee222eecf76d581dca2df6ddb15