Remove unused/wrong TextLayoutResult function
Unused public function which was added for testing.
The function does not do anything usable for Compose
text APIs. The function is now deprecated and should
be removed later.

RelNote: "Deprecated TextLayoutResult/createTextLayoutResult
function. It is an unused public function which was added
for testing. The function does not do anything usable for
Compose text APIs. The function is now deprecated and will
be removed later."

Test: Treehugger

Bug: 184995948
Change-Id: I8041316a98dfd682ca0a8740fc62c8003a0850be


GitOrigin-RevId: 12e1a5b0d872b0e8e9ee8e4a946d09cb02c381e7