Making truth, testing, and utils depend on base modules
Since all of navigation is now in kotlin and the -ktx modules are empty,
we should switch all modules to depend directly on the base module.

Also changed all the modules that use internal-testutils-navigation to
exclude navigation-common base module instead of the ktx.

RelNote: "The Navigation Testing and Navigation Test modules now depend
on the base Navigation modules instead of the empty -ktx modules."
Test: all tests pass
Bug: 184292145

Change-Id: I1b5f5cdb63ae1912cbd9ffb0dbbdc44664fd179a


GitOrigin-RevId: 2a78a49583442afe6872e72d1b822b0f4244ddb7