Fix problem with window having a scroll position.
Fixes: 169804790
Bug: 169874631

It turns out that the problem with b/169874631 was not an error
in getLocationOnScreen(), but just that the window had a scrolled
content. This CL fixes the problem by properly retrieving the
position, including the scroll position of the window.

Also changed cache check to use animation time instead of
drawing time because the view may not draw on every frame
that can have the position changed.

Test: new tests, bubble test, text handles demo
Change-Id: Id907e4dc5d134debc9260fff8c09ced9e7a7a754


GitOrigin-RevId: 991ec6d090dfc0c088803dfa972e9570e6dbd577