Fix flaky DemoTest#testFiltering
The filtering test passes if the back key is pressed before the keyboard is opened.
If the keyboard is opened, then the back key dismisses the keyboard instead of returning
to the root screen. This CL uses Espresso.closeSoftKeyboard() to fix this flake.

Also fixed another flake with Dialogs. We use Espresso.pressBack() to dismiss dialogs,
and then another pressBack() to go back to the root screen. If the last pressBack() is
performed before the dialog is dismissed, we end up stuck on the Demo and don't reach
the root screen. I added a rule.waitForIdle() after Espresso.pressBack() to fix this.

Bug: 185400866
Test: androidx.compose.integration.demos.test.DemoTest#testFiltering
Change-Id: I46948e1f52a19089023b665dc2cb30a571987cf5


GitOrigin-RevId: 90ee3e53da0bc85c10b2e0b8d985e85bbddc2d42