Only invalidate reads of derivedStateOf() if the result changes.
With this change the state object returned by derivedStateOf()
is only invalidated when the value compares != to the
previously value read.

This can significantly narrows the invalidation effect of
changing values read during the calculate lambda of a
derivedStateOf() instance.

Prior to this change, reads of derivedStateOf() where
invalidated whenever any value read by the calculate
lambda was changed even if the result of calculation was the
same.

For example, the state object `derivedStateOf { a > 10 }` will
only invalidate composition when the value of `a > 10` changes
instead of when `a` changes.

With this change `equals()` will now be called for any
value returned by a derivedStateOf() instance during
composition where `equals()` previously would only be called
if the value type was stable and passed as a parameter.

Relnote: N/A
Test: ./gradlew :compose:r:r:tDUT
Change-Id: I05a46499b0b135f1cd27d8ed2f0b0db16ab3123a


GitOrigin-RevId: 80fd1b372600a5597bca8ac9f03a3646590c4d50