Scope WindowRecomposer to a content child
Android can reuse a window decor layout across activity recreations in
the case of dynamically resizing a window. Store the window recomposer
in a content view leaving the decor alone, since the recomposer can be
tied to the local activity lifecycle which may be destroyed in these
configuration changes. This forces the recreation of the recomposer.

Fixes: 184293033
Test: WindowRecomposerTest.windowRecomposerResetsWithContentChild

Change-Id: Ic655335b144fa6adfc857f9838ca00881c689e62


GitOrigin-RevId: f4616ee03d8410d34599ebc25e0341f8902eb056