Migrates Ripple to use RippleDrawable internally
This allows the ripple to be animated on the RenderThread, avoiding jank when the UI thread is under load, such as navigating between screens.

Fixes: b/184842196
Fixes: b/168777351
Bug: b/183019123
Test: MaterialRippleThemeTest
Test: LocalRippleNativeRenderingTest
Test: RippleContainerTest
Relnote: "Ripple has been migrated to use RippleDrawable internally on Android devices. This means that ripple animations will happen on the RenderThread, and so will be smooth even when the UI thread is under load, such as when navigating between screens. This does not change the API surface of ripple, but there may be behavior changes introduced as a result of this change. To aid the migration, `LocalRippleNativeRendering` has been added - provide a value of `false` to this CompositionLocal to fall back to the previous ripple implementation inside the CompositionLocalProvider. This API is temporary and will be removed in the future, so if you run into issues that cause you to use this API, please file a bug: https://issuetracker.google.com/issues/new?component=612128"

Change-Id: I902f8eecaa8e3f55eb4c1bebc86f4f50e2967d0c


GitOrigin-RevId: 152103eb2da716cc91db034ce37bad6b22696b64