Build connection between NavController and Navigator state
Introduce the NavigatorState class. This state is owned
by the NavController, but is specifically set at the
Navigator level. Besides giving the Navigator read access
to the set of destinations on its own back stack, it
also encapsulates the add and pop operations in such
a way as to allow the Navigator to push changes back
to the NavController.

To allow for testing of a Navigator independently
of a NavController, a TestNavigatorState, provides
an alternate implementation specifically for testing
purposes.

Relnote: N/A
Test: updated and existing tests pass
BUG: 80029773
Change-Id: Id2afe4c8d428894826d0391173a162cf8c977328


GitOrigin-RevId: 4ab8db7d2ca746a38bdde67e3ef495e83ddb0aa5