Desktop. Composable Window API. Window. Helper functions and classes to manipulate state
- add LocalWindow. It is needed because we need to know the owner
of the composition in case we need to show a dialog
(a modal dialog needs to know the parent window to be able to disable it)

- UpdatedEffect. Run callback every time the captured state is changed

- ComponentUpdater. Stores applied state and the next time we need to
apply a new state - compares it with previous and updates a component only if this new state is changed.
Similar to androidx.compose.runtime.Updater, but without depending on the current `composer`.

- Data classes to store the window size/position in Dp units

- Safe functions to change the state of Window

- Add ComposeWindow.isFullscreen/isMaximized/isMinimized

Test: ./gradlew jvmTest desktopTest -Pandroidx.compose.multiplatformEnabled=true
Change-Id: Ic6e00106611647c2c7ade6298fb7457c21b6867d