Caches Matrix calls in OwnedLayer implementations
Replaces OwnedLayer#getMatrix() with explicit APIs for mapOffset and mapBounds - this allows the underlying layers to cache their Matrix instances and only update them when the layer properties change, or the layer is moved / resized.

This increases LazyListScrollingBenchmark performance by ~ 4-5%, as we can avoid costly calls to android.graphics.Matrix#nGetValues on every PointerInputFilter#onPointerEvent call.

Bug: b/181869189
Test: LazyListScrollingBenchmark, existing layer tests
Change-Id: I328f87515cb595402dc5e6b01d0cd357ec546bda


GitOrigin-RevId: 7868e12e832e47731f7e819d972311376a0940ad