Adds caching to HitPathTracker nodes
Previously we were translating global positions to local position for each PointerInputChange in each of the three passes, and allocating a PointerEvent for each pass. Instead, since these won't change during dispatching, we can cache this once per pass, and then pass down the modified changes and parent coordinates, so each successive child node only needs to calculate the relative position between them and the last parent we calculated positions for, instead of the root.

Additionally we can cache a map of relevant changes per node, instead of allocating per dispatchChanges call, and only do this work once per dispatch. This allows us to just pass down this map to children, and avoiding needing to mutate InternalPointerEvent#changes, so this also changes InternalPointerEvent#changes to be a val.

These changes lead to reduced allocations and less work, for a ~5% performance improvement in scroll benchmarks.

Bug: b/181869189
Test: LazyListScrollingBenchmark
Change-Id: I114b9c08229eb943c6e1a4b4408beaaffcd3da7a


GitOrigin-RevId: dd162f90ab753288dde57a915453d322f1922a41