Changing the way we determine presubmit builds.
This uses a much better identifier than the previous heuristic.
The key functional difference will be that all local builds will
create configs as if the build were a postsubmit build. You would
need to set BUILD_NUMBER environment variable now, in order to mock
presubmit configs. Since these configs are for test harness ingestion,
and it is not expected that engineers need to debug any difference
locally, this seems acceptable.

Test: Manual comparisson of treehugger build output.
Bug: 182293323
Change-Id: Iaf150c2bb83ebd027e09a0bda59fc3b28c4d2d0a


GitOrigin-RevId: f12e43c54d8ae6d9a79f1597b9ff8b23fe42251b