Ensure TestNavigatorState updates the Lifecycle correctly
The TestNavigatorState serves the role of
"NavController" when building isolated tests
for a Navigator. As such, it owns the logic for
updating and maintaining the state of each
NavBackStackEntry added to it.

As such, the TestNavigatorState now updates
the Lifecycle of each NavBackStackEntry
as it is added and popped from the state's
back stack, mirroring the behavior of
NavController.

Relnote: N/A
Test: new TestNavigatorStateTest
BUG: 80029773
Change-Id: I92b09989a7d9bc63d52747eef40f04d75a43cc0d


GitOrigin-RevId: ff325dad7ef332217b2cb9ae1428d24bb10fbc3b