Remove inline from Modifier.onSizeChanged() and Modifier.onGloballyPositioned()
This allows us to make such modifier skippable. Currently if something triggers recomposition in the function invoking such modifiers those functions will always return a modifier which is not equals to the previous one, causing unnecessary recompositions and remeasurings.

Fixes: 188417479
Fixes: 186109675
Test: ./gradlew bOS
Relnote: Modifier.onSizeChanged() and Modifier.onGloballyPositioned() are not inlined functions anymore
Change-Id: I727f6622174806efb68b63808b322e6bc86e1cf5


GitOrigin-RevId: 35e722280155d3e6135d2a1f4f4ed6e0ef01314c