Allow Navigators to call NavigatorState.pop()
Each Navigator is responsible for coordinating
the external state of each destination with the
NavController by reacting to requests to navigate
to a new destination or pop an existing destination.

However, external events, such as a dialog being
dismissed by the user, means that this coordination
needs to extend both ways. Rather than tie each
Navigator specifically to the NavController, the
Navigator can now call pop() on its NavigatorState
to inform the NavController that the given entry
should be popped.

Test: new NavControllerTest test passes
BUG: 187873799
Change-Id: I43d32c637fcf54000e8fc0af2b5c54fdf8f4810e


GitOrigin-RevId: 511a2f530369d8f225035c08a286701af98ac3ce