Fix for LazyColumn/Row items displayed partially after the scroll
It was the old issue with how we handle not placed children which became often reproducible after introducing slots reuse in LazyColumn. When the layout was not placing its children node remeasuring and redrawing for such node and its children was skipped as this node is not visible. Once this node was becoming visible again we weren't correctly restoring the node's sub-hierarchy. Now we correctly reschedule remeasurings and redrawings for the dirty nodes for the whole subtree.

Relnote: Fixed the bug introduced in beta07 where LazyColumn/Row items were displayed partially after the scroll
Test: InvalidatingNotPlacedChildTest
Bug: 188566058
Change-Id: I8c9ac2df59f3183bf6067dd20092a127f16c9e80


GitOrigin-RevId: a3b566102eb60c3396dfae0dd5ea57e8eb15f0a1