Fixes ripples / indication sometimes getting stuck on a long click
If a long click causes a recomposition and causes the instance of the long click lambda to change, this would cause us to cancel and restart the pointer input scope in Modifier.pointerInput(). This means that we would miss the corresponding up event for that long click, causing no PressInteraction.Release to be emitted, and hence causing ripples / other indication to get 'stuck'.

Now we only cancel and restart this scope if we change from having a long click lambda to not having one (null), and in which case we will now correctly emit PressInteraction.Cancel to clean up any existing long clicks if needed.

Fixes: b/186223077
Test: ClickableTest
Relnote: "Fixed an issue where ripples / other indication would sometimes get stuck on a long click when using Modifier.combinedClickable"
Change-Id: I2298ce564e3940875c3f3525255424da25dc9414


GitOrigin-RevId: 9653864b3c491ed7371ef841bdd292160cbf244e