Do not consume touch events in the Surface.
Consuming the touch events in surface was an overkill. We prohibit any child of the surface to react on the Final pass in the PointerInputEvent processing, causing bugs and unwanted behaviours.

We only need Surface to participate in the pointer input processing so it will hittest positively and, therefore, will block only siblings behind it (e.g. Box with Surface and a Button), which is the only case we really need to support

Fixes: 189411183
Test: added new regression test
Change-Id: I363780e040a47513af8181ded1654c9be47b84fe


GitOrigin-RevId: d4fd71a99aa59c642861d5f5553acce7b7216acf