Update KSP to beta01 and KCT to 1.4.1
* KSP beta01 introduces an api change in function overrides where it can
handle cases when a java method overrides a kotlin property. This also
means the return type of `KSFunction.findOverridee` is a KSDeclaration
instead of KSFunction. We use this when resolving types so I've updated
Room to expect a property declaration as well. It was also affecting
override variance resolution but that code only ever runs for kotlin
sources so it was not necessary to handle property overrides there.

* Adding more tests to java overriding kotlin property revelated another
bug where we were setting an invalid parameter name for property setter
values. (<set-?>). That name is valid in byte code but not in source
code. I've updated JavacVariableElement and KotlinSynthetic to use a
sanitized name instead. We could've done it while generating code but
since many other things refer to these parameters, it is safer to do it
while parsing parameters and hide the detail from the rest of the
processing.

* One of the new tests I've added hit a bug in KSP that revelated after
KSP fixed "java method overrides kotlin property" case. Added a try
catch helper for it:
https://github.com/google/ksp/issues/462

* Previously, when listing methods in a kotlin class, we always listed
setters before getters which made the source code match tests slightly
more difficult. Changed it to keep setter/getter of a property together
in the methods list.

* KSP changed its processor provider interface, updated Room
accordingly. (it now receives 1 container object instead of multiple
parameters)

Bug: 160322705
Test: existing tests + more tests for overrides

Change-Id: I92ec446b59cdeccc37bd91308d1a20acf3fe959a


GitOrigin-RevId: 1ecb11aef121d1970384250c8369a5edaa96dc73