Maintain the scroll position of LazyColumn based on the item's key
In addition to keeping the first visible item index we also store the key of this item. When the user provided custom keys for the items this mechanism allows us to detect when there were items added or removed before our current first visible item and keep this item as the first visible one even given that its index has been changed.
This allows to smartly handle items moves and do not reset the expected by the user scroll position. This behavior is similar to how RecyclerView is handling notifyItemRangeInserted() invocations.

Relnote: Now when you specify the unique keys for LazyColimn/Row items the scroll position will be maintained based on the key, which means if you add/remove items before the current visible item the item with the given key will be kept as the first visible one.
Test: LazyCustomKeysTest
Fixes: 184257857
Change-Id: Id263f45e44fbcd5a6112ef88848da3303705c460


GitOrigin-RevId: 7d26ab7e65a1180eeb6c4bec5a3b018d084f4772