Make swipeable properly handle new anchors while in the middle of the animation
If you have a dynamic content with dymanic anchors that can be changed while in the middle of the animation (e.g. when you open the modal sheet and change the content in the same or next frame). Now the target calculation logic for new anchors accounts for the animation target properly, so we end up in the right anchor.

Fixes: 181593642
Test: added new for swipeable and modal sheet
Change-Id: Ic5d00595fe5611580139aa1a0d87f2e1288507a1


GitOrigin-RevId: f541fc88a477eea1922e5371db2925dedf3a58db