Recreate a content lambda when index/key mapping was changed
It was previously implemented by changing the index baked by MutableState and waiting for the subcomposition to be updated because of this change. But this can only happen on the next frame meaning that the given item would be composed incorrectly on this frame during the measuring. Instead we can just recreate a cached lambda in such rare cases which ensures the subcomposition will we run correctly.

Test: manually and a new test
Fixes: 189966384
Change-Id: I0237b2fb998b11d705e5f3f55d15d7953901855c


GitOrigin-RevId: 4325b171f0e7bb25ebd7409e37183f5d5d6d1ef7