Update isAnimationRunningUpdate test
The test was once written for isAnimationRunning, but that property is
now replaced by isScrollInProgress which has different semantics. In
particular, where `isAnimationRunning` used to be false during a drag,
`isScrollInProgress` is true during a drag.

The actual purpose of the test was to make sure that a fling animation
is interrupted when you put your finger down, so I renamed the method to
reflect this and adjusted the test so it fails if the interrupt doesn't
work properly.

Bug: 189951065
Test: ./gradlew compose:foundation:foundation:cC
Change-Id: Ide714890c456d23d2748f70e6fc18e7382e415eb


GitOrigin-RevId: 425bed504bda3a9626508f7dfbd8e90ad31815a3