[GH] Avoid building playground projects when they are not affected
This PR adds a new Gradle task (`findAffectedModules`) that can report the list of modules that are affected for a given set of files.

That task is later used in the workflow file to check if the playground project is affected by the changes in the RUN and skip building it if it is not affected.

Bug: n/a
Test: ProjectGraphTest, https://github.com/androidx/androidx/actions/runs/939913011

This is an imported pull request from https://github.com/androidx/androidx/pull/197.


Github-Pr-Head-Sha: afb5c6b541b96337c94f99704ce1262a59812c26
GitOrigin-RevId: d55964835dc34977498edccb30ca2413140150fa
Change-Id: I7d676c99a2e60c4712314e6302c861e26b5c3c53


GitOrigin-RevId: 2ca5170a431571076ac988e5367b6d2f8be02c8d