Improve consistency of gradients with transparent colors
On O and later, transparent colors are interpolated
in an elegant manner, ignoring the color channel for
the transparent part. It makes for a smooth transition
to and from transparent values.

On N and earlier, the color channel of transparent
values are also accounted for in the gradient. This
means that Color.Transparent will "blacken" the gradient.

This CL detects transparent colors in a gradient. When
On the ends of the gradient, the proximate color is
used instead of the transparent color, but with an
alpha of 0. When in the middle of a gradient, the
stop is duplicated, with the lower side matching
the previous color, but the alpha 0, and the higher
side matching the next color, but alpha 0. The stop
for both transparent colors are identical.

This works only for transparent colors and not translucent
(0 < alpha < 1). A more complex solution must be used
to approximate gradients with translucent values.

Fixes: 190783390
Test: Added GradientTest to verify color stop logic
Change-Id: I64525313681350e4ed456a5bdb451ffcdde11d52


GitOrigin-RevId: 1ac23550d12fc382ce65f7207c9fd89abe6afa62