Send content change event before scroll event
I observed that View System sends first a content change event and then
a scroll event when scrolling, and furthermore that
ViewRootImpl.handleWindowContentChangedEvent is the only place
that moves the accessibility focus rectangle while scrolling.

I also ought to revisit the 100ms-throttled tree-diff based content
change event system in the future.  But it doesn't look problematic to
send some explicitly associated with scroll observation in addition to
those, as this observer method doesn't require any diffing work to
determine whether a scroll happened.

Test: sendScrollEvent_byStateObservation
Bug: 171509419
Change-Id: I29512b3a19b44f75522a330b247fa5ff5c6f2a1b


GitOrigin-RevId: 83a3f45474698fad162dbd4ec1cd9226ab558270