Fix color of label in error state in text fields
Turns out text field should not display error color of the label when input of the text field is
invalid in case when label plays a role of the placeholder (aka when
text field in not focused and there is no input). Note that material.io
displays text field in error state incorrectly so as our current
implementation of the text field.

Unfortunately, this issue has not been fixed when TextFeildDefaults were
introduced. As such, TextFieldColor.labelColor() function does not have
additional parameter like 'hasInput'. Therefore without an API change
there're two ways I see to fix the issue:

1) showed by this CL which is to change a little bit the meaning of the
`error: Boolean` parameter in the labelColor() function. Previously it
would mean the state of the input of text field. In the new
implemetation this parameter will take into account focus and input of
the text field, and will mean "whether text field should display
label in error state".
The downside is that in the rest of functions of TextFieldColors
interface this parameter will mean the original thing. However, I
noticed that in the rest of the functions this parameter (similar to the
TextField itself) is called 'isError'. But in the labelColor() this
parameter is called 'error'. So this discrepancy can play a significant
role indicating their different meanings.

2) Alternative fix could be to add internal or experimental interaction called InputInteraction to
check inside labelColor() if there is input or not, similar to how we
check if there is a focus. I wasn't sure though if interaction won't be
an overkill for this.

Probably the right thing to do in the future will be to deprecate
labelColor() as it is now and add labelColor() with additional
'hasInput: Boolean' parameter and rename 'error' to 'isError' to fix the
discrepancy with the rest of the functions in the interface.

Bug: 175432263
Test: new unit tests
Relnote: "To align with Material Design specs OutlinedTextField with invalid input stopped using error color for label when label is being used as a placeholder. The latter is true when there is no input text in the text field and the text field is not in focus. With that change the meaning of the `error:Boolean` parameter in `TextFieldColors.labelColor()` function also changed: it will now return `false` even when the input is invalid if the label is being used as a placeholder."
Change-Id: I45f78ef50d31f11ef1776a5a0ed89eb1fac62f93


GitOrigin-RevId: 56e32ef027497860749cdf866df8f83ea986d9be