Fix for stateful components like Text stuck in the wrong state after intrinsics query
It is possible that the component is having some state which is calculated based on the calculated made during the measure pass. And this state is later drawn in the draw modifier.
And then there is a case when the parent of such component needs to be remeasured and this component uses child intrinsics during the measure. How intrinsics work is we essentially run the measure block of a child with not-final constraints and fake measureables assuming that the measure block is pure and side-effect free, which is not true in some cases. What happens is the child component's measure block is rerun with non-final constraints which affected the component state. And then there is no child remeasure with the final constraints as the measuring system thinks that nothing changed for the child and it is already measured to the correct constraints.
The current plan is to force mark the node as needsRemeasure when the intrinsic query happens on it. Sometimes it will cause unnecessary remeasurements but it makes sure such component states are using the correct final constraints/sizes.

Fixes: 191543568
Test: new test in RemeasureWithIntrinsicsTest
Change-Id: I2c43d39756d387b45475c5ce8909a4f861916f97


GitOrigin-RevId: 9c12cdea848c6a0aa15bb0d13bef79499b1d8a40