Disable automatic android component creation.
AGP is working on requiring developers to explicitly pick what
variants should be published.

This fixes the warning:
WARNING:Software Components will not be created automatically for Maven
publishing from Android Gradle Plugin 8.0. To opt-in to the future behavior,
set the Gradle property android.disableAutomaticComponentCreation=true in
the `gradle.properties` file or use the new publishing DSL.

Test: ./gradlew createArchive -> still same results before and after
Change-Id: Icd8e270d74efba89a2cb0a9d372b62913f77121a


GitOrigin-RevId: 86144ee47f83b478ae75029b5f03f2c08bc49368