Align timeline of gestures with timeline of main clock
Invocations of performGesture will now continue the timeline from the
previous invocation of performGesture, while adding the elapsed time on
the main clock to the current gesture time. This prevents that the event
time of the second gesture goes back in time compared to the current
time of the main clock.

Also added a advanceEventTime method to GestureScope so a gesture can
assert more control over the time.

Fix: 192064452
Test: ./gradlew compose:ui:ui-test:cC
Relnote: "Added `GestureScope.advanceEventTime` method to give more
control over the timing of events in a gesture"

Change-Id: Ibf3e2d35b4462863aa0de010cb2d0fe0d10cd3d1


GitOrigin-RevId: 315f533bbfee86549fbca416a0da2db896731fff