Fix LoadState and prevent subsequent InitialPagedList from clearing data
In cases where a user is switching between multiple LivePagedLists, the
placeholder initial value for LiveData, InitialPagedList, is incorrectly
presented to the UI which is a behavior changed from Paging2.

It's still important to have LiveData.value resolve to a non-null value
so this change only affects subsequent submissions of InitialPagedList.

This also fixes deduplication of LoadState from AsyncPagedListDiffer by
properly updating the state in its LoadStateManager

Bug: 192486895
Test: ./gradlew paging:paging-runtime:cC
Change-Id: Ie9ee2cbb8c56b74ab2c240725175858ed7bd868d


GitOrigin-RevId: dce71c319da85935f689ddce8d8a65b54dc8a432