Fix back button handling in Compose
In Android, when a top level focusable item is focused, and the
user clicks back, we exit the app. However in Compose, we first
clear focus and then exit the app when the user presses back again.

This CL removes the need for the extra back press for Compose.

Note: This is an alternate fix to aosp/1753281.
Bug: 192433071
Test: ./gradlew compose:ui:ui:connectedCheck -P android.testInstrumentationRunnerArguments.class=androidx.compose.ui.focus.ComposeViewKeyEventInteropTest
Change-Id: I8a9098a06792668c9e3bafe9485ff2e07b10a8a5


GitOrigin-RevId: 583cec5c1b9c99f841ecd7493666d08ea23cbd50