Update skiko to 0.3.6
A minor update for skiko
fixes incorrect behaviour on Windows.
When the library was loaded multiple
times in multi-classloader environment,
icudtl.dat file was not copied with
the dll file.
Because of it, IDE preview on Windows
was not reloading a preview after changes.

Test: changes in skiko's internals were tested
in skiko test set
(see https://github.com/JetBrains/skiko/pull/131)

Change-Id: I58c95f055aa16e04ceca7a449108f75f706ea66a
Signed-off-by: Alexey Tsvetkov <alexey.tsvetkov@jetbrains.com>


GitOrigin-RevId: 2bf445f3949b3777bfb6ca3ad52a8e1e76c14b88