Refactor trackedTransitions-related data structures
trackedTransitions is currently a ComposeAnimation -> Transition map, which
is a legacy from the old animation clock mechanism in Compose. Now we can
simply use the animationObject property of ComposeAnimation to obtain the
corresponding transition.

In order to reduce the amount of casts from animationObject to Transition,
we also created a named ComposeAnimationSubclass that represents animations
of type ComposeAnimationType.TRANSITION_ANIMATION. Now, trackedTransitions
is a hashset of TransitionComposeAnimation.

This minor refactoring should not impact how the ui-tooling module
commmunicates with Android Studio.

Bug: N/A
Test: ComposeViewAdapterTest andn PreviewAnimationClockTest
Change-Id: I36f1b4e85a0e2244fb550d487e47760d951451f6


GitOrigin-RevId: 058f322ddb6497d81770918bbb26383ae35a1bf9