Make AndroidAccessibilityTest robust to device density
Fixes: 194426666

AndroidAccessibilityTest uses DP for defining dimensions,
which may not align to whole pixels. On some devices that
have fractional DPI settings, tests fail because the test
rounds differently than the actual code.

To solve this problem, the tests were modified to use
pixel dimensions so that no rounding is required.

Test: ran test on Pixel 4 and on 800x480 emulator
Change-Id: I9f21145bd6fbd220c3fcc0ff5f0788019ad99c4f


GitOrigin-RevId: bfbe3bcbacecc42c41cc9abbf5670a8632b03607