Add the Kotlin Version to the build info file
As we continue to develop Compose and other
Kotlin libraries, it will become useful
to be able to easily extract the Kotlin
version for a specific library version.

Right now, our main use case is Compose Compiler
and Runtime.  We can use this to automatically
update DAC with the Kotlin version used for that
version of Compose.

Bug: 194944852
Test: gw createAggregateBuildInfoFiles --rerun-tasks && cat ../../out/dist/androidx_aggregate_build_info.txt
Change-Id: I89ebf689cb6d086d3ba45238db0e86980941e2cb


GitOrigin-RevId: a154005d5ea66b28f8ec97a8836da93ab43c8b06