Having Metalava delete empty removed files
rather than waiting for the executors to complete and then doing it later.

This means we don't need to wait for the WorkerExecutors to complete, which means that Gradle knows that even while GenerateApi is running, it can still start other tasks in the same project.

In practice, this means that compileKotlin can more often run in parallel with generateApi

Bug: 187726157
Test: ./gradlew checkApi
Test: Treehugger runs busytown/androidx.sh

Change-Id: If8572ce902e6cd7d4b2a8c48e85ff8edbfdf1d91


GitOrigin-RevId: f47200b8f8ae407f3466282359230fd14d2078ab