Refactor PageEvent.Insert load state params
PageEvent.Insert will take in two LoadStates param, one for source and
one for mediator, instead of a singel CombinedLoadStates.

This is the second and final part of the effort to consolidate all helper state logic to presenter's side.

One behavior change is that the initial LOCAL refresh where a mediator is present will now send LoadStateUpdate with mediator = IDLE rather than mediator = null.

Bug: 195319029
Test: ./gradlew :paging:paging-common:test
Change-Id: I897bbaedb7d7adb91100bbf73395f67a48c45714


GitOrigin-RevId: 0729c6211f888be129a3e26e0cbc59fedd75cbff