Lower MacrobenchmarkScope min API to 21
Test: MacrobenchmarkScopeTest
Test: ShellUtilsTest
Test: ShellBehaviorTest
Bug: 183129298

Fixes usage of ps -A and pidof.

Started documenting shell version inconsistencies with
ShellBehaviorTest, to be sure we understand which behavior is where,
as testing (local especially) will always be limited, and it's good to
understand these behaviors as best as possible.

Note that low-API level testing has been limited rooted devices so
far, as that's what's available as emulators. Additional fixes likely
needed for user builds of these platform versions.

Change-Id: I7a5243a0570a81b676c4adff2d8267a3b68b61ca


GitOrigin-RevId: eeb028cd24aee126f6f89a59df531304e57ad937