Fix swipeable exception when offset is very close to an anchor.
Bug: 191993377

The logic in the findBounds function that determines when the given offset is an exact match for an anchor accounts for rounding error when comparing the floats. This means that, even when an offset is considered an exact match, the actual offset value might not be exactly equal to the matching anchor value. This breaks the logic in SwipeableState's progress getter, which uses the value returned by findBounds as a key into the anchors map. If that returned value is not exactly equal, the map throws a NoSuchElementException.

This fixes the issue by making findBounds return the anchor that was considered an exact match instead of the offset value itself.

Relnote: "Fix the behavior of SwipeableState in the case where the swipe offset is within a rounding error of an anchor."

Change-Id: I03d39d25bc376314d7197484c2a707498296aa97
Test: tested manually, leaving automated tests until swipeable rewrite


GitOrigin-RevId: 32d335ae09b704ce6e52b3e80d28a1cc8cdae474