Fix concurrent behavior of the state list and map
Prior to this change concurrent access to the result return by
`mutableStateListOf` and `mutableStateMapOf` had undefined
behavior when written to in the global snapshot without
external synchronization. The values for the elements in
the collection not directly affected by the change could appear
to revert to a previous value in some cases as an older copy
of the list or map was used for the change and result of
other changes were discarded.

With this change, all changes are guaranteed to become part
of the state of the list or map by retrying the operation until
the change does not conflict with other concurrent writes. The
repeated loop does not ensure progress and it might cause colliding
operations to take longer on the colliding threads. Use of
a mutable snapshot can be used to ensure there is no looping and
synchronized access to the list or map can be used to ensure
progress.

This change does not ensure changes in the global snapshot are
read/write consistent. Read/write consistency still requires a
snapshot.

However, operations that do not require read/write consistency,
such as the example in b/195886093, will now produce a consistent
result.

Fixes: b/195886093
Test: ./gradlew :compose:r:r:tDUT
Change-Id: Id2384eb415ccc4d4c6abd709080f96ac6b5c6eff


GitOrigin-RevId: eb702ee4af0cfd1abd288c9ef3793922a69f4b11