Update KSP to 1.0.0
I've not gone through resolved issues yet, will do them as followups.
Only updated the tests for https://github.com/google/ksp/issues/159
since the test was working around it and it is now fixed.

Going to stable asap is important for other KSP users (so we don't
conflict with other processors).

Bug: 160322705
Test: existing tests
Change-Id: I6f0684bda98afb04f54be8313cbf041cebc25d32


GitOrigin-RevId: 71c633f602697fc0fee0c0ff20727ffb9e6c8feb