Desktop. Deprecate TestComposeWindow in favor to ImageComposeScene
This CL introduces ImageComposeScene - simplified version of ComposeScene, with drawing only into Image.

It is almost the same as the old TestComposeWindow, but with a different class and method names.

See examples of usage in ImageComposeSceneTest

I "renamed" TestComposeWindow to ImageComposeScene, because its name can confuse people:
1. because we had "Window" in the name, some people thought that it creates the real native window
2. "Test" in the name can mislead about the purpose of the class. The main purpose, of course - is to use it in screenshot tests, but we can also use it in PDF generation, for example.

Also, we introduce:
- renderComposeScene - one-liner function to draw Composable content into an image
- ImageComposeScene.use - similar to AutoClosable.use to disposing created resources. We don't inherit AutoCloseable, because we want to move ImageComposeScene into skikoMain. For now, common code doesn't have Closeable, but in the future it can (https://github.com/Kotlin/kotlinx-io/blob/c5bba0114e33c2afd8e4b1ab214aad94ee416d00/core/commonMain/src/kotlinx/io/Closeable.common.kt#L7), so I renamed `dispose` method to `close`

Also I have fixed a race when we call ComposeScene.render from another thread and use TestCoroutineDispatcher or Dispatcher.Unconfined (see FlushCoroutineDispatcher and ComposeScene).
The race was between:
```
[Swing thread] GlobalSnapshotManager -> sendApplyNotifications -> (ComposeScene.recomposer) awaitWorkAvailable, ..., parentFrameClock.withFrameNanos
```
and
```
[Test thread] ComposeScene.render -> sendApplyNotifications, ..., frameClock.sendFrame
```
In the Swing thread recomposer still don't wait for the frame, but in test thread we already send it.
To make sure that recomposer waiting for the frame, we had `dispatcher.flush`.
But it don't wait for the already running tasks (in our case - for sendApplyNotifications to finish in Swing thread).
In this CL we fix that, adding a synchronization point - we don't exit `flush` if there are still performing tasks.

Test: ./gradlew jvmTest desktopTest -Pandroidx.compose.multiplatformEnabled=true
Test: manual ./gradlew :compose:desktop:desktop:desktop-samples:run1 -Pandroidx.compose.multiplatformEnabled=true

Change-Id: I48d2769fd2eaff53e391371aea9ed37964b6afc3

# Conflicts:
#	compose/ui/ui/src/desktopMain/kotlin/androidx/compose/ui/ComposeScene.desktop.kt
#	compose/ui/ui/src/desktopMain/kotlin/androidx/compose/ui/platform/CoroutineDispatchers.desktop.kt
#	compose/ui/ui/src/desktopMain/kotlin/androidx/compose/ui/platform/FlushCoroutineDispatcher.desktop.kt
#	compose/ui/ui/src/skikoMain/kotlin/androidx/compose/ui/platform/CoroutineDispatchers.skiko.kt