Focus Properties Modifier
The focus properties modifier allows you to set
properties that are applied to a focus target down the
hierarchy. This CL adds a single "canFocus" property
That can be used to control the focusability of an item.
(eg make it focusable only in non-touch mode).

Bug: 163725615
Bug: 189139083
Bug: 155212782
Test: ./gradlew compose:ui:ui:connectedCheck -P android.testInstrumentationRunnerArguments.class=androidx.compose.ui.focus.DeactivatedFocusPropertiesTest
Relnote: "Added FocusProperties modifier that can be used to specify properties for a focus target. For now we have one property that can be used to specify whether the focus target can be focused or not."
Change-Id: If30d43d1040927ed082dec12b3f4fd84d4f143b6


GitOrigin-RevId: 138dc7e62c3860abe03967f591cfb488aaae7b7c