Fix popups crashing when compose view was detached
The crash could happen when the compose view was detached from window.
This caused popup#dismiss due to the popup composition being disposed,
and during this step we were trying to remove the global layout
listener. However, since the compose view had already been detached, the
viewTreeObserver we were referencing was a different one so the listener
was not removed. Further onGlobalLayout calls were causing crashes as we
were calling updateViewLayout on WindowManager for a View that had been
removed.

Fixes: 202870860
Fixes: 200287744
Test: PopupTest
Change-Id: Ifa277b238b131aab6cda538bf773e6698ab9f7a4


GitOrigin-RevId: 93fd5f83b1bf5979b627a4d84326060287be868b