Make proper WindowInsets dispatching optional in VP2
The behavior is now opt-in, because apps relying on stable VP2 will have
the broken behavior and may rely on that.

Fix: 202510973
Fix: 153341849
Test: Added a module with tests for different targetSdkVersions
Relnote: "`ViewPager2` now no longer tries to fix the broken
`WindowInsets` dispatch of old API versions (< 30), because the fix
itself can be harmful to siblings of ViewPager2. The fix is still
available, but has become an opt-in so developers can decide on a case
by case basis. Enable the fix by calling
`WindowInsetsApplier.install(viewPager2))`."

Change-Id: Ic9a85e4d20c35192b783a9436a6f92acec369524


GitOrigin-RevId: e5318a7a92d8aa59166dded878e3d3e41eb9c7af