Fix deadlock when disposing a composition.
Prior to this change, when disposing a composition the compostion
takes its own lock the will remove itself from the recomposer
which requires it take its lock.

When a change is reported to the recomposer it takes its lock then
notifies the composition that is observing the change which takes
its lock.

If these two operations occur simultaniously on separate threads
the threads will deadlock.

Fixes: b/202967533
Test: ./gradlew :composer:r:r:tDUT
Change-Id: Idd68bf2fbe91dd3326ad154651e66914a8d04e97