Add onInitialize() to LayoutNodeWrapper
aosp/1861787 fixed a crash that was caused when LayoutNodeWrappers
were re-used without bring re-initialized. To prevent us from
making a similar mistake in the future, I refactored the code so
that we perform all the initialization code in the onInitialize()
function that will be called whenever the LayoutNodeWrapper is
re-used.

Bug: 203280574
Bug: 203198844
Test: ./gradlew compose:ui:ui:connectedCheck -P android.testInstrumentationRunnerArguments.class=androidx.compose.ui.node.LayoutNodeWrapperInitializationTest
Change-Id: I8951cb547c0e093d8a39593f5f97fb9bb1162e5f


GitOrigin-RevId: 6d0c432fd4ebc751b2815b9d209c57361edbaf81