Fix races when we resize window programmatically or manually
Prevent change window size/position/etc if we recompose in response to OS event
When OS fires event, for example during animation, or when user resizes the window, we called setSize with the reported size. This causes some glitches.

Fixes:
- https://github.com/JetBrains/compose-jb/issues/1040 (on Linux)
- resize jumps on Windows
- "flaky" tests on macOs
- test `WindowStateTest.enter fullscreen before show`

Also fix flaky test on Linux (because of the non-deterministic async event dispatching)

Test: ./gradlew jvmTest desktopTest -Pandroidx.compose.multiplatformEnabled=true
Test: manual resize of the window on each OS
Change-Id: I6a46732a22330eca8a8a521d1e3ac0dd2302c850