Add lint rule for NavGraphBuilder extension in compose scope
When construction your NavGraph within a NavHost in compose, the scope
within the compose builder for the composable function implicitly
references the NavGraphBuilder as 'this' even though you are actually in
a compose scope. This means you can incorrectly call NavGraphBuilder
extension functions within the compose scope and if you attempt to run
your code will cause an IllegalArgumentException.

This change adds a lint rule to prevent calling the composable and
navigation functions within a compose scope since that will always be
incorrect.

Also removed the @Composable annotation from compose samples that were
using it on a NavGraphBuilder extention function.

RelNote: "Calling the `composable` or `navigation` `NavGraphBuilder`
extension functions within a compose scope will now cause a lint error."
Test: added detector test
Bug: 188859857

Change-Id: I5928f5141da7321d464addae51b82df3e085454e


GitOrigin-RevId: e35095d5225cf315e33bc296677744739e60d447