Pass composition locals at the very top level of the Window/Dialog
Currently we pass them not at the top level:
```
composition.setContent {
    ProvideCommonCompositionLocals {
        CompositionLocalProvider(compositionLocalContext) { ... }
    }
}
```
And because of that, compositionLocalContext overrides all locals provided by ProvideCommonCompositionLocals. For example, in this code:
```
application {
  Window {
     Dialog {
     }
  }
}
```
application provides a value for default LocalDensity, and pass it down to Window, and Dialog. But Window and Dialog should provide their own density, as they know on which display they are open.

To fix that, we change order of the providing locals:
```
composition.setContent {
    CompositionLocalProvider(compositionLocalContext) {
       ProvideCommonCompositionLocals { ... }
    }
}
```

Fixes https://github.com/JetBrains/compose-jb/issues/1329
Maybe fixes https://github.com/JetBrains/compose-jb/issues/1376 (need to check)

Partially fixes:
https://github.com/JetBrains/compose-jb/issues/1193
https://github.com/JetBrains/compose-jb/issues/1223
https://github.com/JetBrains/compose-jb/issues/990