[Cherry-pick] Always report the final layout size to the parent during measurement
With the way our measure system works it was possible that some node was first measured with one size and then remeasured with another one if it uses the children sizes.
Imagine the hierarchy like this:
A
-B
--C
Where C has some fixed size and all its parents just use the size of their child.
If after some changes nodes A and C were marked as needs remeasure then we first remeasure A.
A will ask for the size of B and B is not marked as needs remeasure. So A will just use the latest size of the B.
Then we remeasure C and as a result of the remeasure its size has been changed.
This triggers one more remeasure on C, then B and then A.

This means that A was remeasured twice with the wrong size at first. In most cases it causing a bit of extra work, but harmless. However for some stateful layouts like LazyColumn it could be just wrong as it is using its children size in order to calculate the latest scroll position.

In this change I added a logic to detect such use cases and always make sure we remeasured the whole subtree before reporting the measured size to the parent.

Fixes: 179593134
Fixes: 195121516
Test: NodesRemeasuredOnceTest
Change-Id: Ia77ae114ade6a0ce844cebcc52adb7212f8a5c5f