Stop using unsafe cast in SubcomposeLayout
When there was a crash during the subcomposition it is possible that `it.composition` will still be null. Because here I was doing `it.composition!!` the real crash in the composition was then overridden by the new crash in the disposing logic. If we just remove unsafe cast we will make debugging the crashes in the composition simpler.

Fix: 198185880
Test: manually
Change-Id: I5984d9442a781eedd062c2ea2ab1bca57d1a23b1


GitOrigin-RevId: deeb81d4cee1de2097bc196c033f5b4e387e72c5