Make MouseEvent.nativeEvent and KeyEvent.nativeKeyEvent as Any?
Also introduce extensions, which should be used by the users, if they need interop with AWT:
val PointerEvent.awtEvent: java.awt.event.MouseEvent
val KeyEvent.awtEvent: java.awt.event.KeyEvent

We don't yet support non-awt events, as we cast to it internally, but we can do it in the future, when we get rid of this casting.

Make https://github.com/JetBrains/compose-jb/issues/1348 irrelevant.
We can introduce new types in the future.

I don't want to mark it experimental because:
- these types can be used in libraries, it is better if make them backward compatible
- the current PR proposal covers all known cases for our future API
- this API looks pretty simple to write fallback, if we deprecate them someday