Reset pointer IDs after not accepting an input stream
Fixes: 207034744

After compose doesn't consume an ACTION_DOWN MotionEvent, it
won't receive any more events until the next ACTION_DOWN is
received. This means that the next time the pointer arrives,
it should be consided a new pointer so that it does a hit test.

This CL clears the MotionEventAdapter for the pointer that
wasn't accepted when compose doesn't handle it.

Test: manual, new test
Change-Id: If274774c9bf1c923457ec2b1d7caeeba8498de64


GitOrigin-RevId: 951a7a23472e5d77d7d3cf65576038ce6b34bc07