Fix selection of last line
Another attempt to fix https://github.com/JetBrains/compose-jb/issues/1403

Before, bounds was clipped by the outer Box (i.e. by window bounds),
so when we check if the selection handle is out of the box, it returns that it is, but in reality, it is still inside the text.

If we would wrap the Text to scrollable Box, the bounds are correct - it returns bounds of the scrollable content. Before the fix, and after the fix.