Ensure deeplink arg value can be the same as arg name
When matching deep links we specifically check if the given values in the
deep link matches the placeholder that we have in the pattern, and if it
does we provide a default value for that argument if one exists.

This check does not consider the possibility that the value is the same
as the placeholder without the brackets.

We need to ensure that we are actually replacing a placeholder before we
replace the value with anything.

RelNote: "Deeplinks will no longer ignore arguments with values that are
the same as the name of the placeholder."
Test: NavDeepLinkTest
Bug: 207389470

Change-Id: If8017489c9f36da2f20876313356f42ae8bfc208
(cherry picked from commit 40fc519ba8fe0ba87ca39b9f70efefc49cd515cf)
Merged-In:If8017489c9f36da2f20876313356f42ae8bfc208