Don't add selectionMagnifier unless it is actually needed.
The initial SelectionContainer magnifier CL (aosp/1890195) about
doubled the first layout benchmark for the container (b/209754333).
Probably the reason is that the selection magnifier was being added
immediately, and even when there was no active dragging and the
magnifier wasn't being shown. This change only adds the magnifier
when it will actually be shown.

Verified benchmark results locally on a Pixel 3 (max of runs with
1/10/20 children in ns):
Pre-regression: 18389/21368/34316
Regression: 51886/46693/54860
Fix: 15986/17695/28093

Test: No new tests, verified benchmark results manually.
Fixes: b/209754333
Change-Id: Iacbc95443956db7d27f43d175702bd04c112d7f3


GitOrigin-RevId: 4688a46b75ee0a6d031f84bb77df9f6615f9ff8c