Reduce the time taken by snapshot tests
The race condition tests in `SnapshotStateObserver` test were taking
upto 25 seconds to complete. Unit tests should complete in under a
second.

I reduced the iteration counts on these tests reducing the chance they
will catch an error but they should still be useful as they will become
flakey if the error returns and these test are in CI which should
should report them as flakey when if the bug regresses.

Also reduce the a similar count in `CompsoitionTest`.

This takes the time to execute all composition runtime tests from 30
seconds to less than a second.

Fixes: 210480969
Test: ./gradlew :compsoe:r:r:tDUT
Change-Id: Ia5fac3c496ac4d88dea5e27317362d5a09c8f138


GitOrigin-RevId: bd35187f72c311ff10c3e987ed574d66a0f8fd79