Support setSystemGestureExclusionRects
Fixes: 209716068

While it is easy to set a single area to exclude system gestures,
it can be challenging to coordinate setting multiple rectangles
for exclusion.

This adds two new Modifiers specific to Android that make it
easy to exclude system gesture rectangles. The simple one uses
the layout bounds while the one that accepts a lambda uses
the result of the lambda for the exclusion region.

RelNote: "Added excludeFromSystemGesture Modifiers for
easy access to Android's setSystemGestureExclusionRects"

Test: new tests
Change-Id: I46f0768d06c645f23bba4d056448686976e74fca


GitOrigin-RevId: 0d26212202f29ebaad23fd800d17cf4a173b171b