Add Modifiers for WindowInsets
Bug: 213942085

RelNote: "Adds Modifiers for WindowInsets, both for padding
and sizing. This allows some content to extend into the inset area,
and still have primary content stay out of the inset area.
For example, windowInsetsPadding can be used to pad the content
area to avoid areas that may be fully or partially covered."

Test: new tests. manual tests. Tested on L, M, O, P, Q, and R
Change-Id: Id0395142409b981e92f94fb46ebf1c2ec2540048


GitOrigin-RevId: 8e13b9da1fa710b9095a42049d7bd780cc9c16cd