Ensure Lazy layouts recomposition with the same input doesn't cause remeasure
In order to achieve that we have to manually remember the semantics modifier as otherwise it returns the modifier object which is not equals to the previous one causing a remeasure request.
First I tried to only remember the properties lambda param passed into the Modifier factory, but it is not enough to make it work. The reason is that within the properties lambda we create new lambdas for actions like scrollBy and scrollToIndex and because of the new lambdas the equals check is failing.

Test: new tests in LazyListTest and LazyGridTest
Change-Id: I0008a425ba544eafcf57343f7dc20dcb0a01853e


GitOrigin-RevId: d4ac797f1843352053f4332f2c4f19549ac6fbac