Desktop. Remove scheduling events on the next event loop frame
Because of that we sometimes have a clunky unresponsive UI that lives its own live:
- when we resize a heavy undecorated window, it continues to resize, even if we release the mouse (after we upstream this feature https://github.com/JetBrains/androidx/pull/124)
- when we scroll CodeViewer, it continues to scroll for 1-2 seconds. when we scroll fast up/down, it will also be scrolling up/down for some time.

We added it in the past to remove freezes when we scroll heavy lazy lists:
- the user scrolls very fast
- we receive 30-50 events in one frame
- we handle each scroll event (each event can be processed for 50ms), and freeze rendering for some time (for 1-2 seconds)

The better fix for this issue would be either to:
1. handle only the first scroll event, and skip the others in the current frame (this was already implemented here: https://android-review.googlesource.com/c/platform/frameworks/support/+/1837304/33/compose/foundation/foundation/src/commonMain/kotlin/androidx/compose/foundation/gestures/Scrollable.kt#233)
2. or batch multiple scroll events into the one big event (it is planned in https://github.com/JetBrains/compose-jb/issues/1345)

The code inside MouseScrollFilter.kt isn't used anywhere in Compose, it just committed to maintain backward compatibility and to keep codebase in sync with jb-main. It is marked as Experimental in Compose 1.0.0 MPP, so it is probably safe to remove it in 1.2.0. It breaks tests though (not functional itself). Mark it @Ignored for now, will remove when we upstream ImageComposeScene.

Additional fix:

Fix crash when AWT event is sent after the window is disposed (we check isDisposed on every event now). AWT can send event even after calling `window.dipose`

Fixes https://github.com/JetBrains/compose-jb/issues/1448#issuecomment-979466433

Test: ./gradlew jvmTest desktopTest -Pandroidx.compose.multiplatformEnabled=true