Unwrap the context in NavHostFragment
When NavHostFragment was converted to Kotlin, instead of continuing to
use the activity for setting the onBackPressedDispatcher, it was changed
to use the context. In certain situations, such as when using data
binding, you can get a context that is not an onBackPressedDispatcher so
the dispatcher is never set.

We need to unwrapper the context until we find a dispatcher and set it.

RelNote: "The `NavHostFragment` will now properly set the
`OnBackPressedDispatcher` when using viewbinding with nested graphs."
Test: tested in bug sample app
Bug: 214577959

Change-Id: Ifbb512720e49b4cd49b1ba337f5aa6bcebd1a303
(cherry picked from commit 522b48ea3083f758de24d5ff35ea0da90e90db44)
Merged-In:Ifbb512720e49b4cd49b1ba337f5aa6bcebd1a303