Adds a check to icon generation for duplicate icons
On case insensitive filesystems, icons that have the same name but different casing will not correctly be built, as the generated source / class files will end up overwriting each other. In any case it is also just misleading to have icons with the same names but different casing, so adding a check for this will prevent us accidentally adding more icons with this problem in the future.

Fixes: b/216295020
Test: IconProcessorTest
Change-Id: Ieb7ef1a375b4f6d0a5c5faa9e4b2e3aeec12f381


GitOrigin-RevId: dfade4ec15cdaa0150f29d922a85325e5882743e