Decouple ComposeTestRule impl from TestRule
This introduces a new hierarchy of interfaces that do not extend from
TestRule, with a desktop and android implementation of them. The
interfaces are the same as their equivalents that do extend from
TestRule (ComposeTestRule and ComposeContentTestRule), but they cannot
yet replace those equivalents because we haven't decided on the names
yet. They will remain internal for the time being.

All functionality has been extracted from the android and desktop
implementations of the TestRules. The test rules now delegate to an
instance of the new implementation.

Robolectric detection has been adjusted (improved) because we no longer
have the test's Description available when we have to do the detection.
Based on conversations with the Robolectric team, the check has been
improved and now looks at the Build.FINGERPRINT.

This is the first step towards API that allows one to write Compose
tests without depending on TestRule, or JUnit in general. When that API
lands, it will open up the doors to writing tests in commonTest.

Fix: 191899641
Bug: 197851881
Test: ./gradlew compose:ui:ui-test-junit4:cC
Relnote: N/A
Change-Id: I0728cb276eb9345f4fc64c192ce187795cdb940f


GitOrigin-RevId: b8fd1514005fc963af45d8f33aad68d39fa0afb1