Change CoreTextField to use BringIntoViewRequester.
CoreTextField needs to ensure the cursor is in the visible bounds of all
scrollable parents, including the one in the text field itself.
Originally, it tried doing this by going through some of the text
service abstractions and ultimately calling
`View.requestRectangleOnScreen`. However, composable parents also need
to be aware of the request, so the correct way to do this is to use the
relatively recent `BringIntoViewRequester`. I'm not sure why all the
indirection through text services was done, it isn't necessary and
obscures the code, so I just made `CoreTextField` directly call
`BringIntoViewRequester.bringIntoView()`. This helps with b/192043120
but doesn't completely fix it, due to b/216842427.

Bug: b/192043120
Fixes: b/178211874
Test: Tested manually.
Test: Refactored TextFieldDelegate unit tests to use the new code path:
    ./gradlew :compose:f:f:test
Relnote: "`notifyFocusedRect` methods in `TextInputSession` and
    `TextInputService` are now deprecated and won't be called. Use
    `BringIntoViewRequester` instead."

Change-Id: Ia4302c5f6ee79eec30a9f42c149da8775e1ed57e


GitOrigin-RevId: a5c92f31181087df9a6a7100111d4248b94e75d4