Hide top-level dialog from the taskbar
Swing has a special constructor for that (without arguments). It says:
```
A shared, hidden frame will be set as the owner of the dialog.
```

We need exactly this as a default behaviour for top-level Dialog's. This will hide the icon in the taskbar. If the user need to show icon in the taskbar, Window is a better option.

If `SwingUtilities.getSharedOwnerFrame()` was public, I would just use it as a default value in the constructor, but instead we have to create a separate constructor and keep it consistent with JDialog, even if it is unusual API.

The reproducer:
```
import androidx.compose.ui.awt.ComposeDialog
import javax.swing.SwingUtilities

fun main() = SwingUtilities.invokeLater {
    val dialog = ComposeDialog()
    dialog.setSize(800, 600)
    dialog.isVisible = true
}
```

The icon shouldn't be in the taskbar now.