Cherry-pick slider and scrollbar fixes (#183)
* Slider: fix the position of the thumb while dragging

Assume that we have a slide with value range = 0..1, left side at 0px,
and the right side at 100px, and it has a track with left side at 10px
and the right side at 90px. Current implementation maps cursor/touch
position 0px to 0.0 and 100px to 1.0, and the thumb does not follow the mouse cursor. The correct behaviour is to map 10px to 0.0 and 90px to
1.0 because it is the range in which the center of the thumb actually
moves.

Test: ./gradlew :compose:ui:ui:clean :compose:ui:ui:test :compose:ui:ui:desktopTest :compose:material:material:connectedDebugAndroidTest, manual testing
Change-Id: Icd8af3e1e9713838df175eae843ef698f04ad8b6

* Slider: fix position of a thumb while dragging

If mouse moves out of slider container the thumb stops
following mouse cursor. This change allows `rawOffset.value`
to be out of `minPx..maxPx` and coercion is performed just
before computing value.

Test: manual testing (use narrow slider to see what was changed),
./gradlew :compose:material:material:connectedDebugAndroidTest

Change-Id: Ib09c6ffb4aca4c942dc02a96b23718a9389c1a05

* Desktop Scrollbar: fix position of a thumb while dragging

If mouse moves out of scroll area the scrollbar thumb stops
following mouse cursor. This fix separates offset handling
from scrolling logic to avoid unnecessary clamp of offset

Test: ./gradlew :compose:foundation:foundation:desktopTest,
manual testing

Change-Id: Ic38a75e319c88697063bd469370e488e37f97373