Disallow overwriting content with test content
This adds a check in AndroidComposeTest.setContent that makes sure that
the Activity on which the content is set doesn't have content yet. If
there is content already, it is a signal that the tester is probably
unintentionally doing something wrong, so help them by throwing.

Bug: 199631334
Test: ./gradlew bOS
Relnote: "`ComposeContentTestRule.setContent` will now throw an
IllegalStateException if you try to set content when there already is
content."

Change-Id: I888a5054c27d7884110415a812d0ac748be3f869


GitOrigin-RevId: 0fa0145d802d8effe0e833f0fe996dbeec3ab23e