Fix of crash in acessibility code path
It's a blind fix since the issue is not reproducable locally. The logs suggest the parent is getting detached while we're calculating the uncovered semantics nodes. It seems like the activity (and hence compose view) is destroyed but then the accessibility delegate is called. So this fix ensures that node info is not created if we're in destroyed state.

I'll keep the test that was failing flaky and re-enable it later after observing it for a while in CI.
Test: N/A
Bug: 206967867
Bug: 215116019

Change-Id: If9d3a20699056e39b40f27b6a674b9f96cee3420


GitOrigin-RevId: c505dbf74711673fc60871fa6980899a7f70c068