Update showLayoutBounds for Compose even when changed via quick-settings tile.
When the Show Layout Bounds developer option is toggled via quick
settings tile, instead of by going into the Settings activity, the
activity does not get paused and resumed. Currently, we only check the
value of that setting in onResume. This change checks the setting when
the window receives focus as well, which _does_ happen when the
notification shade is dismissed.

Fixes: b/225937688
Test: Tested manually by changing the setting via the quick settings and
observing debug bounds being drawn inside composables.
tile.
Relnote: "The `Show Layout Bounds` setting will now be applied for
    composables immediately after toggling it from the quick
    settings tile, without having to leave and re-enter the activity."

Change-Id: I843d5a7dec3cd7b6b43c0c8af2f928bba51e1532


GitOrigin-RevId: 07ac00dd46a3410dedf1bf7733cf46bb255f862f