[GH] Detect Flow::launchIn in ComposableCoroutineCreationDetector
## Proposed Changes

Compose's `ComposableCoroutineCreationDetector` currently detects disallowed usages of `CoroutineScope::launch` and `CoroutineScope::async` inside composable functions. Another common way to launch a coroutine is `Flow::launchIn`, but the detector doesn't catch it. This adds `launchIn` to the detector.

## Testing

Test: The detector's test code now includes usages of `launchIn`.
I feel that `ComposableCoroutineCreationDetectorTest` is a bit hard to scale up, as line numbers in the expected output will have to be changed after pretty much any change to the test code. I think this test could be improved by breaking it up into smaller tests for each of the cases (plain function, lambda, lambda parameter, etc). I figure this is out of the scope of the PR, but if a maintainer agrees, I'm happy to do that in another PR.

This is an imported pull request from https://github.com/androidx/androidx/pull/258.


Github-Pr-Head-Sha: 18e2400b8d390352d296773ec4c610b3392d3102
GitOrigin-RevId: bc8c86bd575b41876e41128263f50c58b787f48d
Change-Id: Ib52c285d14340272d6cfe7fbb893442744604bf6


GitOrigin-RevId: 0d0772babae4c6121ffceac9e040e4122e26caf4