Switch to dumpsys gfxinfo framestats for detecting launch completion
Fixes: 226179160
Relnote: Overhaul activity launch completion detection to fix 'Unable to read any metrics' exception in startup macrobenchmarks.
Test: TrivialStartupBenchmark
Test: MacrobenchmarkScopeTest#getFrameStats (verified on 23,26,27,28,29)

There are two remaining possible problems that could be causing the
"Unable to read any metrics" exception in startup benchmarks (when the
process is named correctly in the trace, and end of trace marker is
present):

1) `am start -w` doesn't actually wait for first frame to finish rendering
2) activity isn't actually launched / rendered (unlikely, though we've seen this in very specific scenarios in CI)

This change is expected to solve problem 1).

`am start -w` has been historically flakey, and extremely hard to
narrow down problem scenarios, so this CL uses framestats as a
'backup' method to verify launch completion. We now only consider a
launch to be complete when the data for the corresponding first frame
shows up in framestats dump command. Fortunately, framestats is
available back to API 23, the min api of macrobench, though the "flag"
column (which we want to use to detect launch frames) is unreliable
pre API 29 (in both local and CI testing).

Change-Id: Ia517c59c07d9c24e4471e86fb603dca56a6f855f


GitOrigin-RevId: 0ecb70d37c6d0a9c19a2f0c9b289c1d063f9bcf3