Add ComposeUiTest, runComposeUiTest, and friends
This adds the interface ComposeUiTest which contains the same members as
ComposeContentTestRule and ComposeTestRule, but without extending from
TestRule. Since ComposeUiTest is not coupled to JUnit4, it can live in
commonMain to define the API platform independently.

An Android specific variant of ComposeUiTest has also been added,
AndroidComposeUiTest, which provides access to the underlying Activity.
The Desktop implementation, DesktopComposeUiTest, in its turn provides
access to the underlying ComposeScene.

Instead of creating a TestRule through createComposeRule(), new
functions have been introduced to run a test: runComposeUiTest {}. It
sets up the test environment, runs the test (the lambda passed to that
function), and tears down the environment again. This provides
essentially the same functionality as the ComposeTestRules, but without
the dependency on TestRule.

Long term ComposeContentTestRule and ComposeTestRule will extend from
ComposeUiTest, but short term ComposeUiTest is added as separate and
experimental api so we can see if the API shape is the right choice. The
consequence is that for now, the API from the TestRule interfaces is
duplicated in ComposeUiTest, and extension functions on the TestRule
interfaces don't work yet on ComposeUiTest.

Bug: 197851881
Test: ./gradlew bOS
Relnote: "Introduced new experimental, platform independent, test API:
an `interface ComposeUiTest` and a `fun runComposeUiTest(block:
ComposeUiTest.() -> Unit)`, that can be used to run Compose Ui tests
without the need for a TestRule. To run a test without a
ComposeTestRule, pass the test as a lambda to runComposeUiTest, and use
the methods and members in the receiver scope ComposeUiTest, which are
the same ones as in ComposeContentTestRule.

The Android specific `interface AndroidComposeUiTest` and `fun
runAndroidComposeUiTest(block: AndroidComposeUiTest.() -> Unit)` are
added to provide access to the underlying Activity, similar to
AndroidComposeTestRule. For even more control, you can instantiate a
`class AndroidComposeUiTestEnvironment` yourself.

The Desktop implementation is the `class DesktopComposeUiTest`, but no
Desktop specific run functions are offered at the moment.

Migrating a test from a ComposeTestRule to ComposeUiTest can be done
like this (Android example). From:
```
@RunWith(AndroidJUnit4::class)
class MyTest {
    @get:Rule val rule = createComposeRule()
    @Test
    fun test() {
        rule.setContent {
            Text(\"Hello Compose!\")
        }
        rule.onNodeWithText(\"Hello Compose!\").assertExists()
    }
}
```
To:
```
@RunWith(AndroidJUnit4::class)
class MyTest {
    @Test
    @OptIn(ExperimentalTestApi::class)
    fun test() = runComposeUiTest {
        setContent {
            Text(\"Hello Compose!\")
        }
        onNodeWithText(\"Hello Compose!\").assertExists()
    }
}
```

For now, ComposeContentTestRule and ComposeTestRule don't extend from
ComposeUiTest, which means extension functions on ComposeUiTest can't be
called yet on the TestRule interfaces. When ComposeUiTest graduates
to stable API, ComposeContentTestRule and ComposeTestRule will be
changed to extend from ComposeUiTest.
"

Change-Id: Ib4e909d5b4f83a9b3942278e8cb7c86854b6ecaa


GitOrigin-RevId: 11e2362248228175547e29fa37fd58459120f685