Prioritize getRefreshKey if it returns null on invalidate
It is currently documents for the result of `.getRefreshKey()` to be
respected, even if it returns `null` in cases where a non-null
`initialKey` is passed to `Pager`. This resolves an errant use of elivs
operator which incorrectly prioritized `initialKey` in those cases.

Relnote: "The result of `PagingSource.getRefreshKey()` is now correctly
  priotized over `initialKey` in cases where it would return `null, but a
  non-null `initialKey` was set."
Bug: 189492631
Fixes: 230391606
Test: ./gradlew paging:paging-common:test
Change-Id: Ic9542c0062f9a72e5cfa15469bcf0c204a0934bd


GitOrigin-RevId: adc51a26519cb6ed39f88c6da749798c209bad66