Make SelectAll action show the toolbar
Make sure that SelectAll actually selects the text and shows the
toolbar.

Fix a use case where you long press on a blank area (shows Paste/SelectAll menu) and after that you long press on a word. It should
update the menu items (Copy/Paste/Cut/SelectAll) but instead it shows the old Paste/SelectAll items. This is done by finishig the action mode when the long press starts.

This change makes sure that we clear (assign to null) the actionMode when the
action mode is actually finished.

Test: manual testing :(
Bug: 231597471
Change-Id: Ic82a5e3c066bfba9b1de0164f98935d779f09c70


GitOrigin-RevId: 157b3c1e5e62ff54b8db7d8b34b7aa1bebde2bba