Ensure menu items are recalculated when action mode is invalidated
At some point we did an optimisation where  we tried to reuse an action
mode and invalidate it if the action mode already existed.
However, for this to work properly we need to updated the menu items if
they change. We do so by updating the menu in the onPrepareActionMode
call.

Relnote: N/A
Test: new unit test and manual functional testing
Bug: 231597471
Change-Id: I44716c9cef3db78ebb8a49f2d1e88e2d8610f2ad


GitOrigin-RevId: 4b108f6d0b46bee57fd5837f17a62058d14ddf0e