Fix for stretch overscroll clipping the shadow
There is an unwanted behavior in the stretch overscroll effect we have to workaround. When the effect is started it is getting the current RenderNode bounds and clips the content by those bounds. Even if this RenderNode is not configured to do clipping. Or if it clips, but not withing its bounds, but by the outline provided which could have a completely different bounds. That is what happens with our scrolling containers - they all clip by the rect which is larger than the RenderNode bounds in order to not clip the shadows drawn in the cross axis of the scrolling direction. In order to neutralize it we can use similar technique to what we use in those scrolling container clipping by extending the layer size on some predefined constant

Fixes: 214957971
Test: new test in OverScrollTest
Change-Id: Id8886915fa48732e4f67f294fff5bdcc193a12d9


GitOrigin-RevId: 2b02fd36a8c5643c8b8b01623ebe571cb8a0d7bc