docs: Updating Compose Text snippets file
Added a missing snippet (for cleaning text) and updated the
existing snippets with descriptive names (instead of TextSnippet<n>)
as per our updated guidelines.

It looks like adding the downloadable font snippets is more
complicated than I thought, because (per the docs) you need
to do Gradle configuration. I'll try that in a separate CL,
and if I have trouble, I'll ask Ale for help (or to take it over).

Test: ./gradlew :compose:integration-tests:docs-snippets:lintDebug
Bug: 232575723
Change-Id: I3824d9125bda18df0845af55443762588494eae9


GitOrigin-RevId: 195f973f3f5203c39fdf7dfb66cdbac87f5c0837