Noop when FORWARD_DELETE keycode is sent to the end of a text field.
When a hardware keyboard's forward delete key is pressed (typically
called "delete", not "backspace"), `TextFieldKeyInput` searches for the
next character from the cursor and creates an `EditCommand` to delete
it. This also happens for some software keyboards that have a delete
key, such as Samsung's.

We were not handling the case where the `BreakIterator`'s search for the
next character returns a sentinel indicating the end of the string,
which turned into a negative index and caused the crash.

The fix is to explicitly no-op when no character break is found.

However, the whole bug is not fixed, because the only place I could find
where we manually create an `EditCommand` for this creates a
`DeleteSurroundingTextCommand` and the bug report includes stack traces
for `DeleteSurroundingTextInCodePointsCommand` as well. The only way
such a command is created is if the `InputConnection` is sent one, and I
can't figure out how to trigger that yet.

This change also makes the commands in question throw at init when
passed invalid constructor arguments. While this is technically a
behavior change, this behavior was already broken before and it is
unlikely any external code is instantiating these objects.

Bug: b/199919707
Test: Manually tested on emulator with adb to simulate keycode, and
    on a Samsung tablet/soft keyboard.
Test: Added automated test: ./gradlew :compose:f:f:cDAT
Relnote: "Pressing the forward delete key when the cursor is at the end
    of a text field will no longer crash."
Relnote: "`DeleteSurroundingTextCommand` and
    `DeleteSurroundingTextInCodePointsCommand` now require their
    constructor arguments to be non-negative."
Change-Id: Ica8e66d221137a82ac8aaa59372decab096a6ef6


GitOrigin-RevId: d782bbdc9e6e64066eb69ab5ea94640ebbfc534c