Add experimental api to check that we do not nest scrollable containers
This api existed as an internal api for a while as all the usages lived in foundation. However now when we allow users to create their own LazyLayouts they will want to use this assertion in their implementations as well.

Test: n/a
Fixes: 233352497
Relnote: Introduced an experimental api checkScrollableContainerConstraints() which allows to check that we do not nest scrollable containers. You can use it when create your own scrollable lazy layouts via LazyLayout.
Change-Id: Id6fabde2877f20155be3d9642c2e74e0e7d4dab0


GitOrigin-RevId: 02d452019a2cd3d003e7f43ae5c874eb340e9061