Use explicit version for ktlint snapshot
When we use a snapshot version in AndroidX, we should always
use the explicit resolved version that we've downloaded into
prebuilts.

I've also modified the KtLint configuration to force single
version for all of ktlint artifacts to ensure they don't go
out of sync. This is necessary because snapshot artifacts
declare their dependencies as SNAPSHOT too, which means, on
playground, their dependencies might be newer then themselves.

Unfortunately, we have a bunch of deteached configurations
that ignore their dependency rules. I couldn't find an easy
way to make a global configuration resolution rule for ktlint
so had to manually modify breaking ones to account for ktlint.

Bug: 234161459
Bug: 234884534
Test: ./gradlew ktlint
Change-Id: Id54cce37c4fe3e8839367313dcabe6c943c41602


GitOrigin-RevId: 80c9a03de1e39368b6ff9107da31116557a62401