Fix crash in NavHost when fast switching between destinations
There is a currently an error when combining Crossfade with
visibleEntries and fast switching between bottom nav items, where when
going from item 1, to item 2, and back to item 1` before the transition
completes.

Crossfade currently disposes whenever the lastEntry changes, in this
case the old item 1 is not equal to the new item 1` and this causes the
visibleEntries which included item 2 to be disposed when the old item 1
to item 2 transition completes, but the item 2 to item 1` transition is
still going and when that attempts to retrieve item 2, it fails cause it
has already been marked as complete.

We need to make it so that visibleEntries are only marked as complete
when crossfade is being disposed.

RelNote: "Fixed a crash caused by fast switching between bottom
destinations when using the Navigation Compose `NavHost`."
Test: tested in sample app
Bug: 234054916

Change-Id: I3979acdda0b6715fcf96bb1948ad6f5e6ba87388
(cherry picked from commit 95d95a4c54474fc973eebd775ff953f1a2f54c4a)
Merged-In: I3979acdda0b6715fcf96bb1948ad6f5e6ba87388