Stop using warnings as errors
Historically, we used this to make sure we don't have any warnings
sneak into the codebase, but we did not want to force developers from
complying to it locally, so it was default off, but enabled in CI.
Since then, we have added -Pandroidx.validateNoUnrecognizedMessages=true
that can catch any new warnings in output without turning on warnings
into errors. Since having a difference between CI and local is causing
remote cache misses (compiler arguments are different), and we can still
catch these warnings through output validation, this change removes the
use of warnings as errors.

Test: ./gradlew tasks
Change-Id: Ib3250c944c2e5412f004bf193fe02b93cb2b2731


GitOrigin-RevId: 41c58e0ec6622257cbfea73d7487cfd96bcb8fa1