Fix BanConcurrentHashMap lint for Kotlin fully-qualified usages and add tests for more cases
The original issue was that a test for the lint was failing in IMPORT_ALIAS mode. This is actually expected, since the lint deals with import lines. Added a separate test for the case and a comment explaining why the test mode is skipped. In the process found that the part of the lint dealing with fully-qualified usages didn't work in Kotlin, so fixed that and added tests for both Kotlin and Java.

Test: BanConcurrentHashMapTest
Bug: 203124716
Change-Id: I702fc38126eaef19b94c7f24902f026d9dbcdf38


GitOrigin-RevId: ba6d01c96f47a54b1b47a79f55fca455375612a0