Added parameter to `BaselineProfileRule#collectBaselineProfile` to filter output file by package names
Bug: 220146561
Test: ./gradlew :benchmark:integration-tests:macrobenchmark:cC -Pandroid.testInstrumentationRunnerArguments.class=androidx.benchmark.integration.macrobenchmark.BaselineProfileFilterTest

This change adds a new parameter to `BaselineProfileRule#collectBaselineProfile` in order to allow filtering for baseline profile output file and keep only the desired packages. This is useful when creating a baseline profile for a library. Note that, in order to work, the code cannot be obfuscated.

Change-Id: If7338bff7f7a1bde4daeff1af1f684985103012d
Relnote: Added a new parameter to `BaselineProfileRule#collectBaselineProfile` to filter output file by package names


GitOrigin-RevId: a25c7ad45538dca8ebe65bb170ff04605efb8af9