Enable ios for collections
This CL adds ios for collections.

I've changed publishing to copy ios artifacts.
Also changed the KMP plugin to have an ios block
to configure all ios targets. This is not delegating
to the KMP plugin's IOS because it doesn't yet include
arm simulator target.
https://youtrack.jetbrains.com/issue/KT-52312

I've also removed ios32 as there isn't really a recent
enough device that is worth targeting with ios32.

Bug: n/a
Test: ANDROIDX_PROJECTS=KMP ./gradlew bOS --no-configuration-cache
Change-Id: I62350a95fd97526ab21083ca4ede1372a50b3d6e


GitOrigin-RevId: a384058439eac568067513bb0f180ce2304105a3