Create GenerateMetadataTask to create JSON file
By default, this is change does not affect the generated refdocs as metadata insertion is turned off via the `showLibraryMetadata` Boolean in `AndroidXDocsImplPlugin`.

To test, change `showLibraryMetadata` to `true`, and then generate docs before and after this CL.  Check that the generated docs are the same for select Paging pages (paging-compose, paging-runtime, paging-rxjava2, and paging-rxjava3).

Test: presubmit + see above description
Bug: 239095864
Change-Id: I90a4d584557e521a9fd5ae0e11273e4a61340967


GitOrigin-RevId: eb707b469f5e3bd572aa295e5256cc10f9c7582c