Fix merging of leaf nodes with no merging parents
TalkBack has a behavior in AccessibilityNodeInfoUtils.shouldFocusNode()
where direct children of scroll containers automatically merge in all
the speech of their non-screenReaderFocusable children.  This is an
unintuitive/buggy-seeming behavior in the context of Compose where we
promise explicit control of merging.

To suppress this behavior, this patch explicitly marks leaf nodes as
screenReaderFocusable if they are speakable (matching the criteria in
AccessibilityNodeInfoUtils.isSpeakingNode) and none of their parents are
mergeDescendants.

Bug: 239402259
Test: AndroidAccessibilityTest.testScreenReaderFocusable*
Change-Id: I67241fc970d7505adac49205415773309acb2ebe


GitOrigin-RevId: 6dadcd68bd019a41d42e96a2e9c8ed1bab279847