Fix TextField empty line lineHeight (2)
When a line is empty (including empty text) StaticLayout
does not apply the LineHeightSpans.

This causes the line height to be different for an empty
line compared to a non empty line. The issue becomes more
visible within TextField .

This CL stores adjustment values to report the height and the
last line metrics correctly.

For the other solution option please see aosp/2170504

Performance

There are no performance differences for cases where text
is not empty or does  not end with "\n"

The below numbers are for the cases where text is empty or ends with "\n"

+--------+-----------+-----------+------+
|        | Before    | After     | Diff |
+--------+-----------+-----------+------+
| length | ns        | ns        | ns   |
| 0      | 115,835   | 180,161   | 56%  |
| 16     | 395,679   | 484,748   | 23%  |
| 32     | 513,235   | 590,783   | 15%  |
| 64     | 807,133   | 901,896   | 12%  |
| 128    | 1,330,733 | 1,421,371 | 7%   |
| 256    | 2,338,534 | 2,464,810 | 5%   |
+--------+-----------+-----------+------+

Test: Added test
Test: ./gradlew text:text:test
Test: ./gradlew text:text:cAT
Test: ./gradlew compose:ui:ui-text:test
Test: ./gradlew compose:ui:ui-text:cAT
Test: Treehugger

Bug: 236615813

Change-Id: I57f8b6632f569ad06e547cf6ec83df1bc42845c9


GitOrigin-RevId: b7a67dc25a34b254adbaffd02495b5c6d26ee18c