Set text shadow blur to Float.Min when blur = 0
This CL sets the shadow blur to Float.Min when the blur
is defined as 0.

This is required because developers expect the blur=0 to
still show shadow where as platform paint/canvas disables
shadow when blur is 0.

Since currently Text does not use Compose Canvas, and
currently no time travel is possible to fix older platform
versions, the changes are made on Compose Text layer.

Test: ./gradlew text:text:test
Test: ./gradlew text:text:cAT
Test: ./gradlew compose:ui:ui-text:test
Test: ./gradlew compose:ui:ui-text:cAT

Bug: 219552739

Change-Id: Icaa042b4a2e1dc2d74924694432dd46acd1ab31c


GitOrigin-RevId: a18d2ad3c61a3487efaae147388fa24b65c5afcf