Add insets build-in support for m3 navigation components and app bars.
This change adds the insets supports to all navigation components, app bars and scaffold in m3. This allows for the out of the box insets support where the only thing developers would care about is the colors of the status bar icons (depending on the light/dark theme) and the transparency of the status and navigation bars. They would require the Window object to do it, so we cannot automate it from within the components.

Test: Added new compose and activity based tests
Fixes: 183161866
Relnote: Build-in support for Top app bars, drawers, navigation bar and rail has been added. This components, when used separately or with Scaffold will automatically handle insets for developers. Note: This change doesn't add automatic handling of status bar icons and transparency of the status and navigation bars. Please, continue to do it manually to ensure the best edge-to-edge experience.
Change-Id: I7e4e67bd1a84d62bd5ab1eddc7dbed8efdb471d1


GitOrigin-RevId: 848ce53d1abed50947f472861a040a54033202bc