Fix letterSpacing with alignment/ellipsis
When letter spacing, align and ellipsize applied to text,
the ellipsized line is indented wrong. For example for
an LTR text, the last line is indented in a way where the
beginning of the line is less than 0 and the text is cut at
the beginning.

This CL fixes the issue by translating canvas for the
ellipsized line accordingly.

Screenshots: go/compose-letterspacing-align-ellipsis

Test: Added tests
Test: ./gradlew text:text:test
Test: ./gradlew text:text:cAT
Test: ./gradlew compose:ui:ui-text:test
Test: ./gradlew compose:ui:ui-text:cAT

RelNote: N/A

Bug: 228463206
Change-Id: I446445949221027e675cb90d2c33a67c55210aad


GitOrigin-RevId: 39222a2f2ffb2a11efcac1e2d2519bbb2bbc0617