Change return type in CVF lint fix to prevent invalid implicit casting
If a different type is returned, this also changes the name of the generated method by appending the return type in case the same method is used elsewhere in the file expecting a different return type. It is possible that the generated name could still cause a conflict, but it seems less likely.

Test: Added unit tests
Bug: 214389795
Change-Id: I494c49a1d5c1a1c24555d7f54eb618c89cf75636


GitOrigin-RevId: 78e227e639aad4b8941addf8cf1847a3eaeb42dc