Make `remember` lambda parameter `crossinline`
Executing an early return from a `remember` confuses the runtime as it skips
the required end group calls into the composer. This eentually causes a
begin/end imbalence error to be reported as in b/242413202.

This change makes remember's callback a crossinline lambda which prevents
the lambda from being able to early return.

Test: n/a - compiler error
Fixes; 242413202
Relnote:"""BREAKING CHANGE: The parameter to `remember` and
`rememberCoroutineScope` where changed to be `crossinline`. This will report
an error for early returns instead of allowing an early return which will
cause a later internal error to be reported.

This change can potentially lead to new compiler errors to be reported
requiring non-local returns to be removed from the lambdas passed to these
functions."""

Change-Id: Ibea62b31341fb386f47d495ce6bb1708fbfa6c0a


GitOrigin-RevId: cbcd8054849833e441c76dd2bf126b1725c6321f