Lookup CompositionGroups by anchor id
In live mode we do not collect the parameters when go through the
SlotTree to build a component tree.

Before this change the SlotTree was traversed to find a specific
node and collect its parameters.

Recently a find method was added to the runtime which can identify
a node from it's anchor instance. Use that to find the node and
remove the code which was traversing the SlotTable to find a
specific node.

Also add a AnchorMap for generating a unique id for each anchor.
The id is now used in place of the hash value of the anchor which
was passed to the client.

Fixes: 239709530
Test: Existing tests and a new test for the AnchorMap
Change-Id: I634a9c2d6710d76421fa63fd939a536b703a82e1


GitOrigin-RevId: f2c00407e7790db61223e277d45175092e95a856