Fix for small negative overscroll deltas causing the glow effect to dissapear
Small negative deltas that happen when a pointer is mostly stationary cause the glow effect to release - the subsequent positive deltas cause a new overscroll to start and run into a codepath where the delta is so small that the existing glow just instantly dissapears. To workaround this, we now track the cumulative negative delta and only release if it crosses a minimum threshold to prevent accidental releases in the majority of cases.

Test: manual
Fixes: b/242864658
Change-Id: I1428417187084c447890ab19b16577b5c51fadff


GitOrigin-RevId: 4dac4709d864c273b0d3ae4e3d6e418ecc785e0a