Deprecates SmallTopAppBar in favor of TopAppBar
This change deprecates the Material 3 SmallTopAppBar function and
introduces the same functionality as a TopAppBar function.
The change is done to simplify the migration from M2 to M3 and to make
this function more discoverable.

Test: AppBarTest and AppBarScreenshotTest updated
Bug: 226918634
Relnote: "Deprecates the experimental Material 3 SmallTopAppBar function
and introduces and equivalent TopAppBar function. Please migrate your
usage to the new one."

Change-Id: I74404a64a5764ea02895358d2db5dc9bb18c5aba


GitOrigin-RevId: c3bed759a7a041e97b191f6984979800e5c6436c