Fixed race condition when reading state from the global snapshot
There was a race condition when reading state from the global
snapshot that could produce a spurious read error containing the
text "Reading a state that was created after the snapshot was taken..."
when the global snapshot was advanced one thread and being read
on another. Now, if the state record cannot be found, the records
are rescaned with the global snapshot lock to prevent records from
being reused while they are being searched for.

Fixes: b/233779717
Test: ./gradlew :compose:r:r:tDUT

Change-Id: I63e17996494dcda9fc3c4f62ffad735c551e9e64


GitOrigin-RevId: c5ff609874b5f7c19f7dfadcadcb7651d4ee14cd