Support negative spacings in lazy lists and grids
It is a supported use case to pass a negative value to Arrangement.spacedBy() in order to make items overlap. It was mostly working without any modification aside from the one nuance that when we scroll to the top the previous item was not composed anymore when it is visible only because of the negative spacing between items.

Fixes: 244594342
Test: new tests in lazy lists and grids for both regular ones and in tv package
Change-Id: Ibab202f0b9b445c9572379994746d186cd765a16


GitOrigin-RevId: 73175842e122db5f31c7113e6877f097b1638c13