Correctly handle remeasure requests happening while node is not yet placed
It was happening when AnimatedVisibility(which requests new remeasures frequently in order to animate) is used inside SubcomposeLayout and when the child was measured in the placement block (it was written like this in the Scaffold). Our logic was assuming that remeasure requests from a not placed child could be ignored, however we were not rerunning the request once this node is placed. Such children were just stuck in not remeasured state.

Fixes: 247605039
Fixes: 243852341
Fixes: 244063674
Fixes: 244781258
Test: manually and a new test in MeasureInPlacementTest
Change-Id: I385bb15818655587422ab9daf1d2add5bdc61c17


GitOrigin-RevId: 41c4d4c8d7c30965f14cdf2b7de4fbb0042e1cb2