Changed ShellScript to entirely read stdErr file when calling stdErr()
There seems to be a race condition when reading stdErr: the `cat` command may be executed before the actual command now that we don't wait for the first command to finish. This PR returns to the previous behaviour of offering stdErr only after after the full reading. Another solution would be to use `tail -f` to progressively read the stdErr file but not sure we've a use case for that yet.

Bug: 249126534
Change-Id: I8334c0bc4c581a891861bd5a4a8ca9c6dfb1efca
Test: ./gradlew benchmark:benchmark-common:cC


GitOrigin-RevId: 5c1e30529b8b710eba3995a942a300c383e656ae