Change how function instances are compared for skipping
This introduces a new API, changedInstance, on the Composer
and updates the compiler plugin to generate calls to this
new API when it checks function instance parameters for change.

Instances of FunctionInstance cr4eated by using the `::` operator,
for example, return true from equal when the captured state of the
varaible changes. However, this cuases the Composer to not correctly
propagate the value of function instances when the captured state
is updated.

Relnote: N/A
Fixes: b/249050560
Test: ./gradlew :compose:r:r;tDUT :compose:compiler:c-h:i-t:tDUT
Change-Id: Ib52e462a6f11fc8d6d29a778b7966a1d3f54933a


GitOrigin-RevId: 2c479a283867219fc797c822f6a5728e75c77982