Update TextField height calculations according to M3 spec.
M2 behavior: Padding determined the distance between the
top of the text field to the label's baseline, and from
the label baseline to the top of the input field.

M3 behavior: Padding determines the distance between the
top of the text field to the top of the label in focused
state. The input field is placed directly below the label.
Baseline is not considered at all.

Bug: b/250662016
Test: Updated TextFieldTest.kt, updated and cleaned up TextFieldDecorationBoxTest.kt
Change-Id: I85b92ebaee3673fbf49ef941758ab272cf65036f


GitOrigin-RevId: 0769249afe7087e732a17a70d187b6c03a93868f