Fix BitmapSubject.isNotEqualToBitmap pre-emptively stopping before running assertion.
First line of the function compares two bitmaps using `equals` operator. This inequality check always succeeds for different instances, eventually not running `sameAs` assertion for bitmaps. Same bitmaps would not cause an assertion error.

The fix simply gets rid of this unnecessary if check in favor of a null check for the subject.

Test: :compose:ui:ui-text:cAT
Change-Id: I0a3bc56d7491dd405035cb57e1d87fb31b8d4248


GitOrigin-RevId: d2548489aa30adbda4116521d02ee13a8971131d