Simplify multiplatform setup and remove separate library type for KMP libraries.
We don't need to set a property to enable KMP, KMP is enabled if the Kotlin Multiplatform plugin is applied. The Kotlin Multiplatform plugin can only be applied by using the AndroidXPlugin, which is verified. As far as I can tell, at this point the existence of a separate KMP Library type is unnecessary.

By removing this, we no longer encounter errors where developers use the androidXMultiplatform extension, but don't choose KMP Library type, which is confusing and not intuitive to fix.

Bug: 244615383
Test: Existing builds continue to work, and repro case from b/244615383 tested manually.
Change-Id: Ibb3d965a37138e1c6750a136c07c10e549faa220


GitOrigin-RevId: 3b63a24309105e3e1ed8062c4c14a621696491e7