Fix for a crash when Modifier.animateItemPlacement() is set on a child of Lazy grids
It was happening when the item which is visible after the items list change and was not visible before and had an index which is larger than the new items list size.

Fixes: 253195989
Test: new tests in LazyGridAnimateItemPlacementTest
Relnote: The crash with Modifier.animateItemPlacement() and Lazy grids was fixed. It was happening in some conditions when the new items count is smaller then the previous one.
Change-Id: I0bcac889f7916ab032be277ba11fcfb723e56516
(cherry picked from commit c221d62c6aee00a180bf48e776f106396a377a47)
Merged-In: I0bcac889f7916ab032be277ba11fcfb723e56516


GitOrigin-RevId: d29f2a87e3c1d5cb6dfde828400d67b6f161be63