Add BlendMode support to `DrawScope.drawText`
BlendMode coordinates which algorithm is going to be used while painting a new image (source) on a canvas (destination). BlendMode is applied on AndroidTextPaint in the AndroidParagraph class.

BlendMode is only supported for `drawText` calls and it's not a part of TextStyle or SpanStyle.

Relnote: "Added BlendMode parameter to `DrawScope.drawText`, `Paragraph.paint`, and `MultiParagraph.paint` methods to support different blending algorithms when drawing text on Canvas."
Test: :compose:ui:ui-text:cAT
Change-Id: I57508ab06da481f63b4061ecb8ad72c3a733b0a7


GitOrigin-RevId: 3b6bc9fd9a97e4c2d4d50a0e99cd508c61faf4bd