Send accessibility events only if there's a service listening to it
Without a change we only do work like sending events only if touch exploration
is enabled which doesn't cover all services.

In aosp/1938777 we attempted to send events when any accessibility service is enabled. But
it created issues in benchmarks where UIAutomator is used.

In this CL we change how we check if accessibility is enabled: in addition to check AccessibilityManager#isEnabled we're also getting a list of those
services and ensuring it's not empty. UIAutomator doesn't appear in that
list  so nothing changes from UIAutomator perspective.

We also send events by mask: only if there's enabled service that
listens to this type of event we'll send it. This way we won't send more events than
necessary.

Fixes: 212882843
Fixes: 192347141
Test: tested manually with a custom accessibility service that didn't
have touch exploration. Verified that it started to receive the events
that it woudn't before.

Change-Id: I0cded6db9136c93665510523fbce03775e8d092e


GitOrigin-RevId: 8fd26dd3e5bf28bfb9234980013f4c75ec37f575