Migrate M2 ModalDrawer to Swipeable V2
This allows us to test the changes on a stable component and receive feedback from the community.

Test: Existing tests pass
Relnote: Incorporated changes in the Swipeable APIs in `ModalDrawer`. `DrawerState`'s `animateTo` has been replaced by the `open` and `close` methods and the offset is now nullable. Use `requireOffset` to require the offset.
Change-Id: I3de9e361fdd8be7ee83c7946b809b5729d4ef2c4


GitOrigin-RevId: b328bb7473d05ac602056ccb103f313fe0c8d8b3