Support for EmojiCompat in Compose
Add support for EmojiCompat for all Text and TextField in compose

This adds support for EmojiCompat to all Compose based applications.

EmojiCompat configuration must be done outside of Compose, and the global EmojiCompat config will be used.

This change enables automatic configuration of EmojiCompat to compose :ui:ui-text, similar to the configuration in appcompat.

Apps that wish to configure emojicompat monually should continue to disable [EmojiCompatIniitalizer] in their manifest.

Changes:
* Text, BasicText, *TextField, BasicTextField will all now call EmojiCompat.process as needed
* If EmojiCompat loads the font after a Paragraph or MultiParagraph is create, [hasStaleReslovedFonts] will become true

Bug: b/139326806
Test: ./gradlew :com:ui:ui-text:cAT
Relnote: "Added EmojiCompat to Compose"
Change-Id: Ibf6f9f9d37c6280fe1b051269b127f0dfb1d6b6a


GitOrigin-RevId: 358ce0cf2083f6d0e2e2a0897c860a6c1b71b514