Refactor semantics traversal order for more accurate top bars
This change uses 1) the idea of structural significance to determine groups of nodes to visit at a time and 2) row groupings to have more accurate traversal order.

For each node examined, if it is structurally significant, then all of its children must be placed in the final order list before moving on; if the node is not, then all of its children and its children’s children are extracted, then compared with the grouping algorithm. A structurally significant node for now is determined by if it is a container (Surface), scrollable, or a collection.

This fix is mainly focused on addressing top bar ordering issues. The tests added primarily focus on different types of top bars (TopAppBars, Scaffold top bars, top bars with scrolling columns, etc.). New tests were added to `AndroidAccessibilityTest` instead of `SemanticsTests` since the new traversal order is set via ANI properties and needs the ANI provider to test.

In addition to those changes, we no longer need semantic node’s `replacedChildrenSortedByBounds` since all previous sorting has been replaced, so that function, the `sortByBounds` parameter, and the `SemanticsSort` file has been removed.

Tests in `SemanticsTests` that were related to ordering (tests that started with `testChildrenSortedByBounds`) were removed, but new tests with the same coverage were added to `AndroidAccessibilityTest` in their place.

See go/focus-order-changes for more details.

Test: compose/ui/AndroidAccessibilityTest/testCreateAccessibilityNodeInfo_forTraversalOrder_SimpleTopAppBar and others
Fix: b/264431330
Change-Id: I419ae4feec1b319786a6cc32c883d92cb8bac466


GitOrigin-RevId: 74e9c5f21c148a8c68372c1fae290cd733dd0f46