Fix memory leak in Update effects (#325)
If there are animations on the screen, after some time we will encounter performance degradation, and an increased memory consumption.

This fix is similar to the other fix in the past (https://android-review.googlesource.com/c/platform/frameworks/support/+/1398690). We have to remember the lambda instead of creating it each time.

Fixes https://github.com/JetBrains/compose-jb/issues/2455
Fixes https://github.com/JetBrains/compose-jb/issues/1969

The tests are not very deterministic, but it seems better to have them than not to have. The second test fails before the fix, and passes after the fix.
We will see if they will be stable on our CI or not. If they will be flaky, we will tune them, or remove them.

GitOrigin-RevId: 682eeb65d297204a0a3b056af99e267ff1968c49