Removes OverscrollEffect#isEnabled
Instead of layouts configuring whether overscroll is enabled or not based on available scroll space, instead it should be the responsibility of the caller that dispatches events to overscroll, based on information such as ScrollableState#canScrollForward / canScrollBackward. This simplifies the overscroll APIs and also prevents issues such as trying to manually dispatch events to overscroll without remembering to call isEnabled = true first.

Bug: b/255554340
Bug: b/255557085
Test: OverscrollTest
Relnote: "Removed OverscrollEffect#isEnabled. Instead of needing to remember and set this flag, just don't dispatch events to the overscroll effect in cases where you do not want overscroll to show (for example if ScrollableState#canScrollForward/backward both return false)."
Change-Id: I1a4b07627dca47f13d7adb7481fd25097b6eb6a9


GitOrigin-RevId: 322dff0ce2e931fd6c92abe1227fdcd45af72b3c