Rework file output for safety, dont swallow throwables
Fixes: 259567143
Fixes: 227510293
Test: tests in benchmark-common (API 22, 23, 29, 30 31, 33 mostly unrooted)
Test: tests in benchmark-macro (29, 30, 31)
Relnote: "Improve safety of file output writing, which should prevent
output files from silently not being written / appended, especially on
API 21/22."

Fix file output for API 21/22, and avoid swallowing exceptions when
problems occur.

Output.writeFile was previously unpredictably unsafe - on some APIs
and depending on root, it would let the app write to the file, the
shell write to the file or both.

Now we manually expand file access on API 21/22 to allow passing files
between app and shell when necessary. This means we can safely remove
the hack that would allow file append to fail in some situations.

Note that this is extremely difficult to test, as API 21/22 have a
very different directory access pattern, and different defaults for
app-created files (which are globally read/write at the file system
level starting in API 23). Ideally, we would run unrooted devices in
presubmit.

Also added an argument to override thermal throttle sleeps in tests,
since they occur frequently when unrooted on 22, where this was
tested.

Change-Id: If8c444abd9115f214a84e15fbf93d8362d54e58f


GitOrigin-RevId: 5c07c15e0630f0f484172c7cdbee812da12b11f7