Room Gradle Plugin
The Room Gradle Plugin configures the project such that generated schemas that are consumed for auto-migrations and are output of the compile tasks are correctly configured to have reproducible and cacheable builds.

The plugin offers a DSL to configure the base schema location:
```
room {
  schemaDirectory("$projectDir/schemas/")
}
```
The plugin will then configure the Room compiler and the various compile tasks and its backends (javac, KAPT, KSP) to output schema files into flavored folders, i.e. `schemas/flavorOneDebug/com.package.MyDatabase/1.json`. The user then as usual checks-in the files into their repository to be used for validation and auto-migrations.

Bug: 132245929
Test: RoomGradlePluginTest
Change-Id: Ia04df1ec74cd4b230f5a2ea1fb26d6ca89ef944c


GitOrigin-RevId: e8dc18dbaac5515fabcea851ed2c661d836fa198