Update OverscrollBenchmark to allow more time between events.
Fix OverscrollBenchmark to add some time diff between events otherwise a velocity will be wrongly calculated. Because of how the velocity tracker was calculating the velocity, when we were not considering the up events, it was not generating a fling for the OverscrollBenchmark. With the introduction of the up events in the VT calculation, a fling was being generated because the events in this test case were happing too close to one another. The solution is to space the last up event so the VT calculates the velocity without generating a fling.

Fixes: 265351382
Test: N/A
Change-Id: Ibad7d62e0e63dd9ee110a0fce2f85432321f86ed


GitOrigin-RevId: 0eeed411aab7f156bb1bb48b0e8ade57ce71e14c