Fix for scrolling in lazy layouts causing infinity loop
It was happening when the requested offset is larger than the size of the item which was causing our calculated expected scroll position to be the position when the requested item is still not visible. The loop was just trying to scroll by 0 delta over and over again without reaching the item.

Fixes: 259275121
Test: new tests for lists, grids and staggered grids
Change-Id: If18b6c3c549fd50d46dd2bc0e2c747d4bccdb406


GitOrigin-RevId: f729445a60f2a275fc2cee7f2fad37587160435d