Fix node deactivation ordering in SubcomposeLayout
Previously we were just changing active flag to false without calling the apply notification. It could've caused this node which is about to be deactivated to still recompose. For example during such recomposing the whole subhierarchy could be removed because of some other logic (like emitting only if index < itemCount). Removing nodes before deactivation make the reusing of such node useless as there will be nothing to reuse anymore.

Test: SubcomposeLayoutTest
Fixes: 265905168
Change-Id: I146dab8b378d7673a79126aa362fe3ec21993b4a


GitOrigin-RevId: bf87e855903e8b03fba7e816ee7f97a829c3d704