IntermediateMeasurePolicy to support lookahead in SubcomposeLayout
Created an IntermediateMeasurePolicy in SubcomposeLayout such that
animations can be applied in the main pass based on the lookahead
calculations.

The default behavior for IntermeidateMeasurePolicy
is to measure & layout the children _exactly_ as they were
measured and placed in lookahead. This includes:
1. Ensure children was measured in the same phase as they did in
lookahead: whether they were measured during measure phase or
layout phase. They will also be measured with the same constraints
as they did in lookahead
2. Ensure their placement position & placement order are consistent
with lookahead.

This CL also added tracking of placement order for the lookahead pass
to achieve what's described above.

Test: Included

Relnote: "
New SubcomposeLayout API that takes an additional intermediate
measure policy for handling measure/layout logic during
lookahead-based animations.
"

Change-Id: I017d37b9d3c2f890387229bc8cfc6515913c1a3a


GitOrigin-RevId: 782adca53940ee43ec2f64ba02394516a1d35896