Fix potential dead-lock when disposing compositions
When disposing compositions the composition can be
locked then call into the recomposer which locks
the recomposer.

When communicating changes the recomposer would be
locked then lock the composer as it reports changes.

This lack of order in the locks can result in a deadlock.

Changed the recomposer to never be locked when calling
into a composition.

Fixes: 267647309
Test: ./gradlew :compose:r:r:tDUT

Change-Id: I5cd5b2b3388758fef0826af619c680edce07d349


GitOrigin-RevId: 59526cdd95681694b310ae874ec2b1ad198edce0