Lint check for CVF in implicit casts
The check is as described in https://b.corp.google.com/issues/266845827#comment6

Correctly flagging all errors requires finding the expected type of an expression, and the psi library method for this doesn't work for Kotlin. I've added testing for the basic cases (assignment, return, parameter), but this check is definitely not comprehensive for Kotlin.

Fixed existing instances of the issue found by lint, it turns out the `Throwable` cases I'd seen when initially testing (mentioned in bug comment) were all in @RequiresApi classes.

Test: Added unit tests for several cases of valid and invalid casts in both Java and Kotlin.
Bug: 266845827

Change-Id: I62037ca0abccf6bcd70b3bfdaa116ec29a12aff1


GitOrigin-RevId: af75daec4e4d80082aad342a13777f2b11fe925d