[Slider] API Council Feedback
Introduce SliderState which holds all of the information for the current active track for Slider as well as all of the measurements of the components of the Slider. SliderState also holds the gesture and dragging logic.
Updating SliderDefaults.Track to use the SliderState to calculate the values needed to draw the track.

Bug: 269156189
Test: Updating the recomposition tests appropriately with the new SliderState.
Relnote: Added an overload for Slider that is a stateful version of the component. Created SliderState that holds all of the information for the current active track, the measurements of the components of the Slider, and the drag and gesture logic.
Change-Id: I124a5957177062146d7994a115c486a299f1bbb3


GitOrigin-RevId: f88a09b71aefd631588dbfc7302bbdf66e4d6f4e