Workarounds for  b/265438201
This PR introduces 2 changes as workaround for b/265438201. Since
we cannot build for multiple build types at the same time, we need
to handle that gracefully printing warning.

With this PR for both apps and libraries, it's possible to call
`generate<BuildType>BaselineProfiles`. When `mergeIntoMain` is true,
all the variants for that build type will be stored in `main`. When
`mergeIntoMain` is false, it'll work normally creating a profile per
variant of that build type.

Additionally, when calling `generateBaselineProfiles` it will always
trigger the generation for release build type only.

Fixes: 270393496
Fixes: 270392517
Test: ./gradlew :benchmark:integration-tests:baselineprofiles-library-consumer:testBaselineProfilesGeneration :benchmark:integration-tests:baselineprofiles-flavor-consumer:testBaselineProfilesGeneration :benchmark:integration-tests:baselineprofiles-consumer:testBaselineProfilesGeneration :benchmark:benchmark-baseline-profiles-gradle-plugin:test
Change-Id: Id00c4cc6ee496e8a7af792edc6ee02ab995e5ee0


GitOrigin-RevId: 899508e1b199e4095970570a52b677a8af692956