Fixes a11y issues at the M3 date pickers
Fix an issue where a partialy selected day or year is not announced
correctly. We now merge the semantics into the Surface on those
composables in order to have the a11y service read their content even
when the actual Text composable that they hold is not visible (may
happen when scrolling the lists).

Fixes: 269466563
Fixes: 268641948
Test: DatePickerTest, DateRangePickerTest, and Manual testing.
Relnote: N/A
Change-Id: Ia1e46db393d1f941b6e834f640dc8fb19c093bb5


GitOrigin-RevId: 615d56b69dcf31196d5e38100b21d6221c1dd5bc