Redo warning on duplicate files during docs creation
In aosp/2417098 I removed some lines from messages.ignore which were not needed to pass presubmit, but that made the build fail and needed a revert in aosp/2463954.

This is the same changes, but it does not remove the duplicate paths messages.ignore entry for :jvmSourcesJar. It still removes the entry for :compose:foundation:foundation:androidReleaseSourcesJar -- this one is too permissive and would make any warnings from this change meaningless. The build failures were only due to :jvmSourcesJar and :compose:foundation:foundation:androidReleaseSourcesJar doesn't seem to exist (androidReleaseSourcesJar only comes up in messages.ignore on code search and I can't run the task locally), so I think removing it should be okay?

Test: presubmit
Change-Id: Id3e4e728bf0ac495773333da3944c67eb2da63e1


GitOrigin-RevId: adca983629a6bec6874e22cdef96a084f20db096