Refactoring for BaselineProfile gradle plugins
This is an attempt to simplify and refactor the current code.
The AgpPlugin structure hides some of the logic related to agp calls
and tries to keep clean the business logic of the plugin.
Also moved task provider creation in task provider factory methods and
created various useful classes such as ConfigurationManager, R8Utils,
PerVariantConsumerExtensionManager.

Bug: to-add
Test: ./gradlew :benchmark:integration-tests:baselineprofile-library-consumer:testBaselineProfileGeneration :benchmark:integration-tests:baselineprofile-flavor-consumer:testBaselineProfileGeneration :benchmark:integration-tests:baselineprofile-consumer:testBaselineProfileGeneration :benchmark:benchmark-baseline-profile-gradle-plugin:test
Change-Id: I9e6b923c7bacad453f7127cad04887eb13a57d6d


GitOrigin-RevId: d329ada8c1675af0f4d6b75797502d2233eb514e