Reduce allocations by ~25% when parsing XML VectorDrawables
The bulk of this change is to avoid creating new String and FloatArray instances
when possible, and to instead pass start/end offset around to reuse the existing
data buffers.

This change also avoids creating a large array in every call to getFloats() by
reusing a single array we grow as necessary. The original code was creating an
array with a length equal to the length of the string being parsed. A better
heuristic would have been to use at least s.length/2, accounting for the fact
that two consecutive floats need at least one character to act as a separator
so the worst case becomes 2 characters to encode 1 float. We get rid of the
heuristic completely and use a fixed-sized (64) array that grows as needed.

This change also simplifies the parsing code by minimizing state. It should
overall perform fewer operations.

Combining all those changes, parsing paris_30k.xml (17 MiB with >1M points)
goes from 1.6s down to 1.3s, and the number of allocations goes from 16M
down to 12M.

Remaining optimizations:
- Replace String.toFloat() with a custom float parser to avoid a substring
  allocation and to parse faster (we can avoid many edge cases)
- Reduce the number of times we parse the input string. Currently we parse
  it:
  - Once to find the next command
  - Once to find the next start of a float
  - Once to parse the actual float
  This means we iterate over the input string roughly 3 times. We could
  instead merge all those steps and parse once
- Further reduce complexity/allocations by emitting PathNode instances
  as we parse instead of going through intermediate steps. It's a little
  less necessary now that allocations have been greatly reduced but we
  could this way get rid of the growing float array by just emitting
  one float at a time in the current node

Test: PathParserTest
Test: ImageVectorTest
Test: VectorTest

Change-Id: I12c2654f39cab0e3250b4be41edd99f32cf18d91


GitOrigin-RevId: de31e126b92760888d0eeaacf389df03ee7a7852