Changed consumer plugin tests to use android test modules as producer
The consumer test module previously used as baseline profile dependency
an ad-hoc module that would provide the same configurations expected by
the producer and provide artifacts the same way. This required
additional maintenance and was less precise, as a test, than relying
on a module with the producer plugin.

This PR disables the `connected<variant>AndroidTest` and manually sets
the `resultsDir` to point to a tmp folder with a fake `test-results.pb`
generated ad-ho for each test. This allows to test the full interaction
between the different modules, serving as a better test than the
previous one.

Test: ./gradlew :benchmark:integration-tests:baselineprofile-library-consumer:testBaselineProfileGeneration :benchmark:integration-tests:baselineprofile-flavor-consumer:testBaselineProfileGeneration :benchmark:integration-tests:baselineprofile-consumer:testBaselineProfileGeneration :benchmark:benchmark-baseline-profile-gradle-plugin:test
Bug: 272775801
Change-Id: I7b33d5803f3eaab5cc62020fce0d54660ecd29eb


GitOrigin-RevId: 5abec16fcb7901659e8282e1d3044918249338cf