Fix `endToMarker()` when ending node groups.
`endToMarker()` is called when a composable function calls
an inline function that has a non-local return. If the non-local
return would return passed the `endNode()` call of an inline
composable function `endToMarker()` was supposed to call `end()`
of the composer with `true` but it was unconditionally calling
it with `false` causing the runtime to get confused as to when
to generate a node.

Also if the `currentMarker()` was called when recomposing existing
content but the `endToMaker()` was called with new content the
groups in the new content would not correctly be closed if the
current marker is in the direct parent group of the new content.

Fixes: 264467571
Fixes: 274889428
Test: ./gradlew :compose:r:r:tDUT
(cherry picked from https://android-review.googlesource.com/q/commit:d71d980734a752e516140dd3ac031af5bab86aa3)
(cherry picked from https://android-review.googlesource.com/q/commit:0a19c2606f19d4ae6775de31a85c958f0f3597c8)
Merged-In: Ibe7063cf22f4bff6eda5bde05c37c1e665c09167
Change-Id: Ibe7063cf22f4bff6eda5bde05c37c1e665c09167


GitOrigin-RevId: 928d9ae402bbcb6cd27263f5bbf6232f0eacc391