Support partial results and profile updates
This PR changes how the CollectBaselineProfileTask works to support
partial baseline profile generation, when the test runner arg
`android.testInstrumentationRunnerArguments.class` is detected.
When a full result is detected, the output dir is deleted and recreated
in order to delete any existing profile. When a partial result is
detected, only the regenerated profile is overwritten.

Test: ./gradlew :benchmark:integration-tests:baselineprofile-library-consumer:testBaselineProfileGeneration :benchmark:integration-tests:baselineprofile-flavor-consumer:testBaselineProfileGeneration :benchmark:integration-tests:baselineprofile-consumer:testBaselineProfileGeneration :benchmark:benchmark-baseline-profile-gradle-plugin:test
Fixes: 274836264
Change-Id: I61465785382eba6e8ca3b6436e9809959b4def6d


GitOrigin-RevId: e34384a86216c33f5cf9fa70cbe83d0e39401833