Changed AgpVersionAttr to BaselineProfileGradlePluginAgpVersionAttr
This PR changes AgpVersionAttr to
BaselineProfileGradlePluginAgpVersionAttr so that the AgpVersionAttr
disambiguation rule is not used. The rule throws an exception if the
agp version is not the same. In this case the baseline profile gradle
plugin is using its own version of the value, calculated from the
AndroidPluginVersion instance in the project. In AGP this seems to be
coming from a Version class. This may cause some misalignment in the
string version of this value. We cannot directly use the version class
as it seems not to be exposed. We could also remove the attribute but
I think there it's better to provide gradle an additional attribute for
matching configurations here.

Fixes: 275564821
Test: ./gradlew :benchmark:integration-tests:baselineprofile-library-consumer:testBaselineProfileGeneration :benchmark:integration-tests:baselineprofile-flavor-consumer:testBaselineProfileGeneration :benchmark:integration-tests:baselineprofile-consumer:testBaselineProfileGeneration :benchmark:benchmark-baseline-profile-gradle-plugin:test
Change-Id: Ib5ed727bf5ccb862a28146e46d38f79ed0389788


GitOrigin-RevId: 18f268ae05ecc1745f0c476484d828162ec9e610