Fix forceMeasureSubtree leaking into lookahead pass
Prior to this fix, forceMeasureSubtree does both
measure and lookahead measure for children, even
when the forceMeasureSubtree was initially invoked
by a non-lookahead measure call (i.e.
layoutState == LayoutState.Measuring). As a result,
some of the lookahead invalidations were not properly
handled.

This change isolates the non-lookahead forceMeasureSubtree
to only non-lookahead measurements, and defer any
lookahead invalidations in the subtree to afterwards.

Test: Added
Bug: 274614438
Change-Id: I6da8de7bb6d790ab9c95c6e1c8ebce79166bef77


GitOrigin-RevId: 3bc37d9c407a4e34e2bc8d555e7f067c56fe094b