Fix FileTestIO not to eagerly create files
The expectation in these tests is for the file not exist but current
implementation of FileTestIO was creating them eagerly. That eager
creation fails when the parent directory doesn't exist, which fails
the test.

I assume this didn't fail before because we happen to run
OKIO tests before this (which created the directory, so it passed,
even thouogh it wasn't the test's intention. The test wanted a non
existing directory.).

This is a temporary fix for this particular case. We need to do a
cleanup to ensure these two IO test abstractions behave the same and
also clean after themselves. Otherwise, we'll keep getting this kind
of test flakes based on the order they are run.

Fixes: 276970642
Bug: 276983736
Test: androidx.datastore.core.SingleProcessDataStoreTest#testWriteToNonExistentDir
Change-Id: I0ab8c6d92308c5917f5f2269ecb7040143f30800


GitOrigin-RevId: d4bb5b14ec30b8527ae604a3b091f67d26af9481