[M3][ModalBottomSheet] Move ModalBottomSheet to android only implementation.
ModalBottomSheet tying to screen size, anchors and Windowing are very explicitly android based, and as we continue to iterate on Popup logic the desktop version falls behind. ModalBottomSheet is also not recommended for desktop use.

This also lays foundation for bottom sheet anchors which can handle edge to edge.

Test: Unit tests WAI
RelNote: ModalBottomSheet moved to android only. ModalBottomSheet is not recommended for desktop use, and the functionality is not actively maintained.
Change-Id: Ib37784ddf10366ce34ecbcab3bcf15cc534b26fc


GitOrigin-RevId: 6b2c350aafd7b6a12fc195b9f521c60fc2a0f442