[M2] Move thresholds into components.
We are moving the density dependency to the component level.
This applies to the following components: BottomDrawer, ModalBottomSheetLayout, BottomSheetScaffold, Switch, ModalDrawer

Relnote: We are moving the density dependency to the component level. This applies to the following components: BottomDrawer, ModalBottomSheetLayout, BottomSheetScaffold, Switch, ModalDrawer. Please use the new overload provided where density is a parameter.

Test: Existing tests
Fixes: 277048962
Fixes: 277049443
Fixes: 277049955
Fixes: 277195326
Change-Id: I8fbd8011386bbaa46e3a1492cbaa9766d8f24024


GitOrigin-RevId: c9cbf5cfca3a14ed24fdc8e574da174aed56afe7