Extract logic that decides whether or not a SemanticsNode (and its AccessibilityNodeInfo) will be `screenReaderFocusable`.
Test: Existing tests (no-op refactoring)

When `currentSemanticsNodes` are retrieved, their order is sorted and saved in an `idToBefore` map. Then when ANIs are created, each one is individually populated: `screenReaderFocusable` is determined based on other properties, and then the `traversalBefore` and `traversalAfter` properties are also set. But we eventually want to sort nodes only if they are either screenReaderFocusable or a traversalGroup. So the logic that determines whether an ANI is `screenReaderFocusable` or not must be taken out of the `populateAccessibilityNodeInfoProperties`, which is what this CL does.

Change-Id: I1a18e11bcfb84a9ecccf1121b97a4a7c218af10e


GitOrigin-RevId: 0094ab0395af0c3befc92cca185042cbb47c83cc