Default onTextLayout to null
The new Text modifier refactor checks if onTextLayout
is null in order to decide on which modifier to use. If
there is no text layout callback, it can use a faster version.
See TextUsingModifier.kt:80

This small change could see a big impact to real world apps
by putting them on a faster path with no code change.

Bug: 278172768
Test: Existing tests
Relnote: NA
Change-Id: Ia3d479b689da3948639ed3004fef2dd28d04ff76


GitOrigin-RevId: dcf56889c843c032539ba754eb5b63cd1f6ddfb6