Add warning for opaque Unit arguments used as keys
The remember, disposableEffect, rememberSaveable, launchedEffect, and
produceState composables all accept keys that, when changed, trigger
old values to be invalidated and disposed. While it is sometimes valid
to pass in `Unit` as an argument here, it is easy to do so accidentally
by passing a Unit-function as a key argument or by writing a more
complex expression that always resolves to `Unit`.

This CL adds a lint inspection to find Unit expressions used as keys to
these relevant functions, and issue a warning. The recommended fix is
to pass `Unit` explicitly and move the expression before the block.
Passing the `Unit` type explicitly will not be flagged by this
inspection.

Test: OpaqueUnitKeyDetectorTest
Change-Id: I71d0373981658717caf81e9a7d890cfd2e398322


GitOrigin-RevId: 497a2446b6974ff373717f248c222658a4919236