Ensure FragmentNavigator does not destroy entries early
There is a case where if you navigate to a destination and the fragment
that is assocated with the destination does a navigate with popUpTo in
its onResume() lifecycle callback, the FragmentNavigator incorrectly
destroys the entry before the associated fragment is destroyed. The
reason for this is the timing between fragment lifecycle callbacks and
Lifecycle. The onResume() lifecycle callback happens immediately before
the `ON_RESUME` Lifecycle event and once the lifecycle listeners in
FragmentNavigator get the `ON_RESUME` event, the state of the entry has
changed and gets destroyed.

Instead of always marking entries as complete when we get the end
Lifecycle events (ON_RESUME and ON_DESTROY) from the fragment, we should
check whether the back stack contains the entry or not.

RelNote: "Fixed an issue in Navigation with fragments where removing a
fragment via navigate with popUpTo in its `onResume()` lifecycle
callback would cause an `IllegalStateException`."
Test: added test
Bug: 279644470
(cherry picked from https://android-review.googlesource.com/q/commit:320aa4ca44fc50e7131fd7571321f0c23aa9ef87)
Merged-In: I21884c268cf213caacabe19db38fd49a1ac829fa
Change-Id: I21884c268cf213caacabe19db38fd49a1ac829fa