Make sure content is not disposed before being reused in lazy layouts
Before this change sometimes we were first disposing the item content when the associated item is being removed, and only after that we reuse the subcomposition. This makes the reuse less efficient as there are no layoutnodes to reuse left. This change makes sure layoutnodes are not released until we decide on do we need to keep the slot for reuse.

Test: new test in LazyLayoutTest
Change-Id: I39665e4cc74909a943c0aec7061294f5db8217b8


GitOrigin-RevId: 3a8622835744a82365735133c63a283fcc01a3cb