Fix bug in FocusInvalidationManager
This fixes a bug in FocusInvalidationManager. We have checks for FocusEvent
and FocusTarget nodes where we check if the node is un-attached before we
invalidate it, but we didn't have this check for FocusProperties.

This CL also adds messages to the other such assertions in the code.
(The exception encountered by the playstore app was thrown from one of the
DelegatableNode functions, but since it was an inline function, it appeared
as if the assertion originated from FocusInvalidationManager. These messages
should help make it easier to find the root cause in the future.

Bug: 280941088
Test: Confirmed that the Playstore app does not crash
Change-Id: Ie2f04195627bfd6756d1c7a64ffc325257bc0447


GitOrigin-RevId: e78cbb1dff61c95d7d1ff5f42bbb19c09cd0ec17