1.5.0-beta03 merge. Fix CoreText.kt (#664)
- Apply changes from jb-main between the base commit 6022301db806601f282c53b8cbb5a981923a1589 and the pre-merge commit e961a81ec0e1c252f33b39801e7e621514289342:

  - 77edab66 (this commit was a squash, and its message is a mess)

The `pointerInput(Unit)` is restarted every recomposition, because we create `mouseSelectionObserver` every time.

It can cause selection issues. We can accidentaly not have issues (they depend on how `mouseDragGestureDetector` is used inside TextField), but it can return in the future, so better to write these modifiers in a proper way.

GitOrigin-RevId: 240287b05fd58a85ab14d3c5c4f1f19e86194852