Fix scrolling when we change LazyListState (#769)
* Fix scrolling when we change LazyListState

Fixes https://github.com/JetBrains/compose-multiplatform/issues/3551

When the new state is passed to LazyList, it goes down to `Modifier.pointerScrollable` as `controller: ScrollableState`, and wraps to `rememberUpdatedState`, which is always the same instance. Because it is the same instance, MouseWheelScrollableElement.update is never called.

Passing it is as a value now, as MouseWheelScrollableElement doesn't need a state.

* Remember ScrollingLogic

It seems there is test that checks equality of the modifier across compositions. Probably for optimization.

Remembering scrolling logic seems logical here, and shouldn't add overhead.

In 1.6 it was fully rewritten to modifier node:

https://github.com/androidx/androidx/blob/a7e6bdf524be186baca08e9ddf7272b834a14bd2/compose/foundation/foundation/src/commonMain/kotlin/androidx/compose/foundation/gestures/Scrollable.kt#L276

GitOrigin-RevId: db4f869fecc110f32037dc28c768e55d01727bc4