Fix entry not marked complete after view is destroyed
As part of a fix for b/279644470 in aosp/2566210, when the fragment view is destroyed, we would mark the transitioning (outgoing) entry as complete only if the entry has also been removed from backstack. However, in the simple navigate() case where the outgoing entry is not popped and remains in the backstack, this entry is not marked complete and ends up in the visibleEntries even though the view has been destroyed.

Now the backstack check in ON_DESTROY is no longer necessary to fix the original b/279644470 bug -- the view being destroyed should be the right signal for completing entries regardless of whether transitioning effects are used. So this check is removed.

Test: ./gradlew navigation:navigation-fragment:cC
Test: ./gradlew navigation:navigation-runtime:cC
Bug: 288520638
(cherry picked from https://android-review.googlesource.com/q/commit:f61da31809827bfac89af629c88ad9b8423e88b6)
Merged-In: I5caa9af1b5bd7084e76d7daf9515f7430bf2489d
Change-Id: I5caa9af1b5bd7084e76d7daf9515f7430bf2489d