Clean up WrapJsComposableLambdaLowering from decoys-related checks and properties
This fixes the compilation for k/js and k/wasm targets. WrapJsComposableLambdaLowering was requiring some decoy-related inputs, which were not provided (for k/wasm).
Providing them doesn't make a lot of sense, because they're not going to be used anyway (decoys to be removed completely).

It's a follow-up to https://android-review.googlesource.com/c/platform/frameworks/support/+/3039572

Test: tested against Compose Multiplatform testsuite
Change-Id: Ib44f7023fdd19d3259f6b8d86f1dcc16957be3c4