Adopt common Drag&Drop API for multiplatform
This change is targeting to unblock implementations of Drag&Drop APIs on platforms where transfer session might be initiated only by the system itself (iOS, Web).

To do that `dragAndDropSource` now implements default start detector that not only allows delegating it to the system, but also simplifies the usage of it. Specifying custom behavior is still possible by passing a new `detectDragStart` parameter (not available on some platforms). Inside the scope of this detected there is an ability to request starting the transfer.

In addition of entry point changes in `compose:foundation`, there are important `compose:ui` API changes:
- Requesting of start transaction now limited by callback scope.
- Splitting `DragAndDropModifierNode` into `DragAndDropSourceModifierNode` and `DragAndDropTargetModifierNode`
- Removing from `DragAndDrop*ModifierNode` APIs that shouldn't be exposed outside of ui module
- `DragAndDropManager` now public with `@InternalComposeUiApi` annotation. It's the same `DragAndDropNode` that is used now as API for `DragAndDropManager` (replacement for removed APIs from interfaces)

Also these changes fix a few scenarios (even on Android):
- Cleanup `interestedTargets` (b/326945380)
- Sometimes drop to the target was just ignored. It was due to recreation of delegating node (was easily reproducible even in demo/sample)
- `DragAndDropNode.contains` check used incorrect size in case of applied padding modifier

Relnote: `Modifier.dragAndDropSource` was adopted to multiplatform - now it implements default start detector that not only allows delegating it to the system, but also simplifies the usage of it.
Fixes: 326945380
Bug: 330481832
Test: existing tests + manually
Change-Id: I0389dd1406de633206650eec717d1ae778734911