Avoid running MissingSerializableAnnotation on binary files
If the KClass we are checking for `@Serializable` is part of binary,
it is not possible for the developer to add the annotation anyway. So we
should stick to only linting classes that are part of the source.

Test: ./gradlew :navigation:navigation-runtime:lintRelease
Bug: 372274443
(cherry picked from https://android-review.googlesource.com/q/commit:ec7a60be7a61d8da4006ad536cabaf3169a1d6a2)
Merged-In: Ic67299007b0f0c70f63a1ab4e43075cd75374b86
Change-Id: Ic67299007b0f0c70f63a1ab4e43075cd75374b86