Add conditional check inside NavHost PredictiveBackHandler
There is no way to guarantee that we will get a composition if a state
change and back callback get triggered in the same frame. This can
result in triggering the `PredictiveBackHandler`'s callback even though
it should actually be disabled.

The cleaner options of fixing this are to change the
`PredictiveBackHandler`'s `enabled` parameter to a Boolean producing
lambda that can be executed internally and retrieve the current value,
but that needs a new API. (see 379152996#comment2)

In the meantime, we just need to add the same check inside of the
`PredictiveBackHandler`'s lambda.

RelNote: "Fixed an issue where `NavHost` could throw an exception
inside of the `PredictiveBackHandler` if the back stack is popped down
to 1 entry and a system back are triggered in the same frame."
Test: added testDisabledInSameFramePredictiveBack
Bug: 375343407
(cherry picked from https://android-review.googlesource.com/q/commit:f724837c72814a08c4177004ae3e8f380eea8367)
Merged-In: I1a4297a10f2d0da8c8477644b9250c61e4792923
Change-Id: I1a4297a10f2d0da8c8477644b9250c61e4792923