Set the default version to 9999.0.0-SNAPSHOT, pin lifecycle in Compose UI to 2.8.4
These are preparation to build Compose 1.8.0-alpha01 from `jb-main` branch.

1. Set the default version to 9999.0.0-SNAPSHOT. It is needed for pinning, more in the code comment.

**Alternative**
Move versioning from CI to `libraryversions.toml`. It will contain all next versions we need to publish, and they are always greater than the pinned versions. This requires some work and I didn't plan it in the short term.

**Disadvantage**
We can accidentally publish it to the dev repo. But users won't see it in the suggestions because it is `-SNAPSHOT`.

This is also true for the alternative, so we have this disadvantage anyway.

2. Pin lifecycle to `2.8.4` in `jb-main` branch

**Alternatives**
- Change jetbrains.publication.version.LIFECYCLE=2.8.4 in gradle.properties (an option, but there is no verification that it uses the correct API)
- Don't pin, depend on the project version, which should be 2.9.0-alpha01 (but we need to publish Lifecycle, and CI for it is WIP)
- Create `release/1.8.0-alpha01`, pin there (doesn't work, because the current release process requires version increasing in the main branch after we create a `release` branch, to maintain ordering of release/dev builds)

**Disadvantages**
- "Go to definition" goes to the 2.8.4 library sources instead of the project sources.
- Can't use/test new changes in Lifecycle

Proper pinning guidelines proposal is WIP at the moment, we can discuss it later, and for now, agree on a working solution for 1.8.0-alpha01.

## Testing
1. Build
```
./gradlew publishComposeJbToMavenLocal -Pjetbrains.publication.version.COMPOSE=1.8.0-test1 -Pjetbrains.publication.libraries=COMPOSE
```
2. Check module files
3. Sync and run the version on the desktop template:
- lifecycle should be 2.8.4
- runtime should be the same that was built