Set the default version to 9999.0.0-SNAPSHOT, pin all dependencies (#1696)
These are preparation to build Compose 1.8.0-alpha01 from `jb-main`
branch.

1. Set the default version to 9999.0.0-SNAPSHOT. It is needed for
pinning, more in the code comment.

**Alternative**
Move versioning from CI to `libraryversions.toml`. It will contain all
next versions we need to publish, and they are always greater than the
pinned versions. This requires some work and I didn't plan it in the
short term.

**Disadvantage**
We can accidentally publish it to the dev repo. But users won't see it
in the suggestions because it is `-SNAPSHOT`.

This is also true for the alternative, so we have this disadvantage
anyway.

2. Pin all dependencies to be able to not release them.

Not pinned dependencies:
- collection (new API needed)
- core (new lib core-uri)
- shapes (no stable version)

## Testing
1. Build
```
./gradlew publishComposeJbToMavenLocal
```
2. Check module files
3. Sync and run the version on the desktop template
4. run1, runDemo